% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{add_reference}
\alias{add_reference}
\title{Add a reference plan to a set of plans}
\usage{
add_reference(plans, ref_plan, name = NULL)
}
\arguments{
\item{plans}{a \code{redist_plans} object}

\item{ref_plan}{an integer vector containing the reference plan. It will be
renumbered to 1..\code{ndists}.}

\item{name}{a human-readable name for the reference plan. Defaults to the
name of \code{ref_plan}.}
}
\value{
a modified \code{redist_plans} object containing the reference plan
}
\description{
This function facilitates comparing an existing (i.e., non-simulated)
redistricting plan to a set of simulated plans.
}
\concept{analyze}
