% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varinfo.R
\name{compare_plans}
\alias{compare_plans}
\title{Make a comparison between two sets of plans}
\usage{
compare_plans(plans, set1, set2, thresh = 0.1)
}
\arguments{
\item{plans}{a \link{redist_plans} object}

\item{set1, set2}{\code{\link[dplyr:dplyr_data_masking]{<data-masking>}} indexing vectors
for the plan draws to compare. Must be mutually exclusive.}

\item{thresh}{the value to threshold the eigenvector at in determining the
relevant set of precincts for comparison.}
}
\value{
A list with the following elements:
\item{eigen1}{A numeric vector containing the first eigenvector of
\code{p1 - p2}, where \code{p1} and \code{p2} are the co-occurrence matrices
for \code{set1} and \code{set2}, respectively.}
\item{eigen2}{A numeric vector containing the first eigenvector of
\code{p2 - p1}, where \code{p1} and \code{p2} are the co-occurrence matrices
for \code{set1} and \code{set2}, respectively.}
\item{group_1a, group_1b}{Lists of precincts. Compared to \code{set2}, in the
\code{set1} plans these precincts were much more likely to be in separate
districts. Computed by thresholding \code{eigen1} at \code{thresh}.}
\item{group_2a, group_2b}{Lists of precincts. Compared to \code{set1}, in the
\code{set2} plans these precincts were much more likely to be in separate
districts. Computed by thresholding \code{eigen2} at \code{thresh}.}
\item{cooccur_sep_1}{The difference in the average co-occurrence of precincts
in \code{group_1a} and \code{group_1b} between \code{set2} and \code{set1}.
Higher indicates better separation.}
\item{cooccur_sep_2}{The difference in the average co-occurrence of precincts
in \code{group_2a} and \code{group_2b} between \code{set1} and \code{set2}.
Higher indicates better separation.}
}
\description{
This function provides one way to identify the structural differences between
two sets of redistricting plans. It operates by computing the precinct
co-occurrence matrix (a symmetric matrix where the i,j-th entry is the
fraction of plans where precinct i and j are in the same district) for each
set, and then computing the first eigenvalue of the difference in these two
matrices (in each direction). These eigenvalues identify the important parts
of the map.
}
\details{
The co-occurence matrices are regularized with a \eqn{Beta(1/ndists, 1-1/ndists)}
prior, which is useful for when either \code{set1} or \code{set2} is small.
}
\concept{analyze}
