% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_calculations.R
\name{compute_log_target_density}
\alias{compute_log_target_density}
\title{Computes the unnormalized log target density for plans}
\usage{
compute_log_target_density(
  map,
  plans,
  counties = NULL,
  sizes_matrix = NULL,
  compactness = 1L,
  constraints = list(),
  num_threads = 0,
  pop_temper = 0L
)
}
\arguments{
\item{plans}{Either a \code{redist_plans} object or a matrix or vector of plan labels}

\item{sizes_matrix}{A matrix of region sizes. The matrix must have
the same number of columns as \code{plans}, \code{ndists} rows, and each column
must sum to \code{ndists}.}

\item{num_threads}{The number of threads used for computing
the target densities.}
}
\value{
A vector of the unnormalized target density of each plan
}
\description{
Computes the unnormalized log target density for the standard redist
target measure. The density for a plan  can be decomposed into a plan-term
(if relevant) and the procuct over regions of the regions compactness and
its score
}
