% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_calculations.R
\name{compute_log_target_density_by_region}
\alias{compute_log_target_density_by_region}
\title{Computes the unnormalized log target density contribution for each region in a plan}
\usage{
compute_log_target_density_by_region(
  map,
  plans,
  counties = NULL,
  sizes_matrix = NULL,
  compactness = 1L,
  constraints = list(),
  num_threads = 0,
  pop_temper = 0L
)
}
\arguments{
\item{num_threads}{The number of threads used for computing
the target densities.}

\item{plan_matrix}{A matrix of 1-indexed plans}

\item{init_region_sizes_mat}{A matrix of region sizes. The matrix must have
the same number of columns as \code{plan_matrix}, \code{ndists} rows, and each column
must sum to \code{ndists}.}
}
\value{
A matrix of number of regions by plans of the unnormalized target density of each region in a plan
}
\description{
Computes the unnormalized log target density contribution for each region in
a plan. This amounts to the log number of spanning trees minus the score. Note
that if the target measure has a plan-wide component then the sum of the log
densities over regions is not neccesarily equal to the log density of an entire plan.
}
