% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{east_tx}
\alias{east_tx}
\title{East Texas Precinct File}
\format{
sf tibble containing columns for useful data related to the
redistricting process, subsetted from real data in Florida, and sf geometry column.
\describe{
\item{\code{county}}{The FIPS code for the county containing the precinct.}
\item{\code{id}}{A unique FIPS-like identifier for the precinct.}
\item{\code{geometry}}{The sf geometry column containing the geographic information.}
\item{\code{cd}}{The current congressional district assignment.}
\item{\code{pop}}{The total population of the precinct, according to the 2010 Census.}
\item{\code{black}}{The non-Hispanic Black population of the precinct.}
\item{\code{hisp}}{The Hispanic population (of any race) of the precinct.}
\item{\code{white}}{The non-Hispanic white population of the precinct.}
\item{\code{vap}}{The voting-age population of the precinct.}
\item{\code{bvap}}{The Black voting-age population of the precinct.}
\item{\code{hvap}}{The Hispanic voting-age population of the precinct.}
\item{\code{wvap}}{The white voting-age population of the precinct.}
\item{\code{dem}}{Number of votes for Hillary Clinton in the precinct in 2016.}
\item{\code{gop}}{Number of votes for Donald Trump in the precinct in 2016.}
}
}
\usage{
data(east_tx)
}
\description{
This data contains geographic and demographic information on 3,138 precincts
from the eastern portion of the state of Texas.
}
\examples{
\dontrun{
data(east_tx)
}
}
\concept{data}
