% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{EPSG}
\alias{EPSG}
\title{EPSG Table}
\format{
named list containing EPSG codes for each U.S. state. Codes are
indexed by state abbreviations.
}
\usage{
data("EPSG")
}
\description{
This data contains NAD83 (HARN) EPSG codes for every U.S. state.
Since \code{gredist} uses projected geometries, it is often a good idea to use
projections tailored to a particular state, rather than, for example, a
Mercator projection. Use these codes along with \code{\link[sf:st_transform]{sf::st_transform()}} to
project your shapefiles nicely.
}
\examples{
data(EPSG)
EPSG$WA # 2855
}
\concept{data}
