% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_smc.R
\name{est_norm_unbiased}
\alias{est_norm_unbiased}
\title{Unbiased Estimates the log of normalizing constant of plans sampled with \code{redist_smc}}
\usage{
est_norm_unbiased(map, plans)
}
\arguments{
\item{map}{A \code{\link[=redist_map]{redist_map()}} object.}

\item{plans}{A \code{redist_plans} object generated using \verb{Redist 5.1} or later.}
}
\value{
An estimate of the normalizing constant from each run
}
\description{
Returns an unbiased estimate of the log normalizing constant for plans
sampled using \code{redist_smc}. Requires that the plans were sampled with
\code{control = list(est_norm_unbiased = TRUE)}, which draws one extra particle
per step and discards it so that the attempt counts give an unbiased
estimate.
}
