% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fl25_adj}
\alias{fl25_adj}
\title{Florida 25 Precinct File}
\format{
A list storing the adjacency graph for the 25-precinct subset of Florida.
}
\description{
This data set contains the 25-precinct shapefile and related data for each precinct.
All possible partitions of the 25 precincts into three contiguous
congressional districts are stored in \code{\link{fl25_enum}}, and the
corresponding adjacency graph is stored in \code{\link{fl25_adj}}.
}
\examples{
data(fl25_adj)

}
\references{
Fifield, Benjamin, Michael Higgins, Kosuke Imai and Alexander Tarr.
(2016) "A New Automated Redistricting Simulator Using Markov Chain Monte Carlo."
Working Paper. Available at
\url{http://imai.princeton.edu/research/files/redist.pdf}.
}
\concept{data}
