% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fl25_enum}
\alias{fl25_enum}
\title{All Partitions of 25 Precincts into 3 Congressional Districts
(No Population Constraint)}
\format{
A list with two entries:
\describe{
\item{\code{plans}}{A matrix containing every partition of the 25 precincts
into three contiguous congressional districts, with no population constraint.}
\item{\code{pop_dev}}{A vector containing the maximum population deviation
across the three districts for each plan.}
}
}
\usage{
data("fl25_enum")
}
\description{
This data set contains demographic and geographic information about 25
contiguous precincts in the state of Florida. The data lists all possible
partitions of the 25 precincts into three contiguous congressional districts.
The 25-precinct shapefile may be found in \code{\link{fl25}}
}
\examples{
data(fl25_enum)

}
\references{
Fifield, Benjamin, Michael Higgins, Kosuke Imai and Alexander Tarr.
(2016) "A New Automated Redistricting Simulator Using Markov Chain Monte Carlo."
Working Paper. Available at
\url{http://imai.princeton.edu/research/files/redist.pdf}.

Massey, Douglas and Nancy Denton. (1987) "The Dimensions of Social Segregation".
Social Forces.
}
\concept{data}
