% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_canonical_plan_labelling}
\alias{get_canonical_plan_labelling}
\title{Get canonically relabeled plans matrix}
\usage{
get_canonical_plan_labelling(plans_mat, num_regions, num_threads)
}
\arguments{
\item{plans_mat}{A matrix of 1-indexed plans}

\item{num_regions}{The number of regions in the plan}

\item{num_threads}{The number of threads to use. Defaults to number of machine threads.}
}
\value{
A matrix of canonically labelled plans
}
\description{
Given a matrix of 1-indexed plans (or partial plans) this function
returns a new plans matrix with all the plans labeled canonically.
The canonical labelling of a plan is the one where the region of the
first vertex gets mapped to 1, the region of the next smallest vertex
in a different region than the first gets mapped to 2, and so on. This
is guaranteed to result in the same labelling for any plan where the
region ids have been permuted.
}
\details{
Modifications
\itemize{
\item None
}
}
\keyword{internal}
