% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_map.R
\name{get_existing}
\alias{get_existing}
\title{Extract the existing district assignment from a \code{redist_map} object}
\usage{
get_existing(x)
}
\arguments{
\item{x}{the \code{redist_map} object}
}
\value{
an integer vector of district numbers
}
\description{
Extract the existing district assignment from a \code{redist_map} object
}
\concept{prepare}
