% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_map.R
\name{get_existing_seats}
\alias{get_existing_seats}
\title{Extract the existing seat counts from a \code{redist_map} object}
\usage{
get_existing_seats(x)
}
\arguments{
\item{x}{the \code{redist_map} object}
}
\value{
an integer vector of the seats in each district
}
\description{
Extract the existing seat counts from a \code{redist_map} object
}
\concept{prepare}
