% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_map.R
\name{get_graph}
\alias{get_graph}
\title{Extract the adjacency graph from a \code{redist_map} object}
\usage{
get_graph(x)
}
\arguments{
\item{x}{the \code{redist_map} object}
}
\value{
a zero-indexed adjacency list
}
\description{
Extract the adjacency graph from a \code{redist_map} object
}
\concept{prepare}
