% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{get_k_step_ancestors}
\alias{get_k_step_ancestors}
\title{Get k-step Ancestors of particles}
\usage{
get_k_step_ancestors(parent_mat, steps_back = NULL, start_col = NULL)
}
\arguments{
\item{parent_mat}{Ancestor matrix where entry \link{i,j} equals the index of the
parent of particle i after step j}

\item{steps_back}{How many steps back we should find the ancestors of (so
parents are 1 step ancestors). Defaults to going all the way back to the
original ancestors}

\item{start_col}{Which column to start at. Defaults to the last column}
}
\value{
indices of k-step ancestors for particles at iteration \code{start_col}
}
\description{
Tells you for a given index what its original ancestor was.
}
