% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_alg_helpers.R
\name{get_map_parameters}
\alias{get_map_parameters}
\title{Takes a map and possibly a county label and returns map parameters for redist algorithms}
\usage{
get_map_parameters(map, counties = NULL)
}
\arguments{
\item{map}{A \code{redist_map}}

\item{counties}{Either a numeric or factor vector mapping each vertex in the
graph to a county.}
}
\value{
A list with 0-indexed adjancency list, county label vector,
population bounds vector, and population vector.
}
\description{
Takes a \code{redist_map} and possibly a county label vector and after checking
the inputs are valid it returns a list with the adjacency map, county vector,
and population variables
}
