% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_plans.R
\name{get_mh_acceptance_rate}
\alias{get_mh_acceptance_rate}
\title{Extract the Metropolis Hastings Acceptance Rate}
\usage{
get_mh_acceptance_rate(plans)
}
\arguments{
\item{plans}{the \code{redist_plans} object}
}
\value{
a numeric acceptance rate
}
\description{
Extract the Metropolis Hastings Acceptance Rate
}
\concept{simulate}
