% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_plans.R
\name{get_nseats_matrix}
\alias{get_nseats_matrix}
\title{Extract the matrix of the number of seats for each district from a redistricting simulation}
\usage{
get_nseats_matrix(x)
}
\arguments{
\item{x}{the \code{redist_plans} object}

\item{...}{ignored}
}
\value{
matrix
}
\description{
For \code{redist_plans} object with \code{M} plans, \code{ndists} districts and \code{S} total
seats this returns a \code{ndists} by \code{M} matrix where each column sums to \code{S}.
}
\concept{analyze}
