% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{get_original_ancestors_mat}
\alias{get_original_ancestors_mat}
\title{Get Original Ancestor Matrix of particles}
\usage{
get_original_ancestors_mat(parent_mat)
}
\arguments{
\item{parent_mat}{Ancestor matrix where entry \link{i,j} equals the index of the
parent of particle i after step j}
}
\value{
indices of originals ancestors for particles at every iteration. So
entry \link{s,j} is the original ancestor of particle j after step s
}
\description{
Gets a matrix of the original ancestors (ie first splits) of the particles
at each steo,
}
