% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_map.R
\name{get_pop_tol}
\alias{get_pop_tol}
\title{Extract the population tolerance from a \code{redist_map} object}
\usage{
get_pop_tol(map)
}
\arguments{
\item{map}{the \code{redist_map} object}
}
\value{
a single numeric value, the population tolerance
}
\description{
Extract the population tolerance from a \code{redist_map} object
}
\concept{prepare}
