% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_map.R
\name{get_ref_plan_and_seats}
\alias{get_ref_plan_and_seats}
\title{Extract the reference plan from a \code{redist_map} object. If the map was
previously subsetted the districts in this plan will be labelled 1 to \code{ndists}}
\usage{
get_ref_plan_and_seats(x)
}
\arguments{
\item{x}{the \code{redist_map} object}
}
\value{
\code{NULL} if no reference plan exists. Else returns a list with
\itemize{
\item \code{ref_plan} - a vector of precinct assignments if a reference plan exists, if not
returns \code{NULL}
\item \code{ref_seats} - a vector of seats for each district
}
}
\description{
Extract the reference plan from a \code{redist_map} object. If the map was
previously subsetted the districts in this plan will be labelled 1 to \code{ndists}
}
