% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_plans.R
\name{get_sampling_info}
\alias{get_sampling_info}
\title{Extract the sampling information from a redistricting simulation}
\usage{
get_sampling_info(plans)
}
\arguments{
\item{plans}{the \code{redist_plans} object}
}
\value{
a list of parameters and information about the sampling problem.
}
\description{
Extract the sampling information from a redistricting simulation
}
\concept{simulate}
