% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_splitting.R
\name{get_subtree_from_root}
\alias{get_subtree_from_root}
\title{Build New Tree from Vertex}
\usage{
get_subtree_from_root(tree_adj_list, root)
}
\arguments{
\item{tree_adj_list}{The adjacency list of the tree. It should be zero indexed and
directed.}

\item{root}{The 0-indexed vertex to start building the new tree from.}
}
\value{
The subtree created by traversing the tree only starting from \code{root}
}
\description{
Takes a tree stored in a 0-indexed, directed graph form and returns the subtree
created by only traversing the tree from \code{root}.
}
