% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_map.R
\name{get_target}
\alias{get_target}
\title{Extract the target district population from a \code{redist_map} object}
\usage{
get_target(x)
}
\arguments{
\item{x}{the \code{redist_map} object}
}
\value{
a single numeric value, the target population
}
\description{
Extract the target district population from a \code{redist_map} object
}
\concept{prepare}
