% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumpart.R
\name{gredist.calc.frontier.size}
\alias{gredist.calc.frontier.size}
\title{Calculate Frontier Size}
\usage{
gredist.calc.frontier.size(ordered_path)
}
\arguments{
\item{ordered_path}{path to ordered path created by gredist.prep.enumpart}
}
\value{
List, four objects
\itemize{
\item \code{max} numeric, maximum frontier size
\item \code{average} numeric, average frontier size
\item \code{average_sq} numeric, average((frontier size)^2)
\item \code{sequence} numeric vector, lists out all sizes for every frontier
}
}
\description{
Calculate Frontier Size
}
\examples{
\dontrun{
data(fl25)
adj <- gredist.adjacency(fl25)
gredist.prep.enumpart(adj, "unordered", "ordered")
gredist.calc.frontier.size("ordered")
}
}
\concept{enumerate}
