% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_majmin.R
\name{gredist.plot.majmin}
\alias{gredist.plot.majmin}
\title{Majority Minority Plots}
\usage{
gredist.plot.majmin(grouppercent, type = "hist", title = "")
}
\arguments{
\item{grouppercent}{output from gredist.group.percent}

\item{type}{string in 'hist', 'toptwo', or 'box'}

\item{title}{ggplot title}
}
\value{
ggplot
}
\description{
Majority Minority Plots
}
\concept{plot}
