% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindex.R
\name{gredist.sink.plan}
\alias{gredist.sink.plan}
\title{Sink Plans to 1:ndists}
\usage{
gredist.sink.plan(plan)
}
\arguments{
\item{plan}{vector of assignments, required.}
}
\value{
A vector with an ID that corresponds from 1:ndists, and attribute \code{n} indicating the number of districts.
}
\description{
Takes a plan and renumbers it to be from 1:ndists
}
\examples{
data(fl25_enum)
plan <- fl25_enum$plans[, 5118]
# Subset based on something:
plan <- plan[plan != 2]
plan <- vctrs::vec_group_id(plan)
# Now plan can be used with redist_flip()
plan

}
\concept{prepare}
