% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{imp_confint}
\alias{imp_confint}
\title{Confidence Intervals for Importance Sampling Estimates}
\usage{
imp_confint(x, conf = 0.95, .data = cur_plans())
}
\arguments{
\item{x}{\code{\link[dplyr:dplyr_data_masking]{<data-masking>}} the vector to
build importance sampling confidence intervals for.}

\item{conf}{The confidence level for the intervals.}

\item{.data}{a \code{\link{redist_plans}} object}
}
\value{
A tibble with three columns: \code{X}, \code{X_lower}, and
  \code{X_upper}, where \code{X} is the name of the vector of interest,
  containing the mean and confidence interval. When used inside
  \code{\link[dplyr:summarise]{summarize()}} this will create three columns in the
  output data.
}
\description{
Builds a confidence interval for the mean of a vector of interest,
given importance sampling weights.
}
\concept{analyze}
