% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_alg_helpers.R
\name{infer_plan_nseats}
\alias{infer_plan_nseats}
\title{Attempts to infer the number of seats for each district in a plan}
\usage{
infer_plan_nseats(
  plans,
  total_seats,
  precint_pops,
  lower_pop_bound,
  upper_pop_bound,
  num_threads = 0L
)
}
\arguments{
\item{plans}{Either a \link{redist_plans} object or a 1-indexed plans matrix}

\item{total_seats}{The total number of seats in the plan}

\item{precint_pops}{A vector of precinct population assignments with same
number of rows as the plans matrix.}

\item{lower_pop_bound}{The minimum population bounds for a single seat}

\item{upper_pop_bound}{The maximum population bounds for a single seat}

\item{num_threads}{The number of threads to use while calculating. Defaults to
maximum availible on the machine.}
}
\value{
A matrix of the number of seats for each region in \code{plans}
}
\description{
Attempts to guess the number of seats assigned to each region in a plans
matrix.
}
