% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{iowa}
\alias{iowa}
\title{Iowa County File}
\format{
sf tibble containing columns for useful data related to the
  redistricting process
\describe{
\item{\code{fips}}{The FIPS code for the county.}
\item{\code{cd_2010}}{The 2010 congressional district assignments.}
\item{\code{pop}}{The total population of the precinct, according to the 2010 Census.}
\item{\code{white}}{The non-Hispanic white population of the precinct.}
\item{\code{black}}{The non-Hispanic Black population of the precinct.}
\item{\code{hisp}}{The Hispanic population (of any race) of the precinct.}
\item{\code{vap}}{The voting-age population of the precinct.}
\item{\code{wvap}}{The white voting-age population of the precinct.}
\item{\code{bvap}}{The Black voting-age population of the precinct.}
\item{\code{hvap}}{The Hispanic voting-age population of the precinct.}
\item{\code{tot_08}}{Number of total votes for president in the county in 2008.}
\item{\code{dem_08}}{Number of votes for Barack Obama in 2008.}
\item{\code{rep_08}}{Number of votes for John McCain in 2008.}
\item{\code{region}}{The 28E agency regions for counties.}
\item{\code{geometry}}{The sf geometry column containing the geographic information.}
}
}
\usage{
data("iowa")
}
\description{
This data contains geographic and demographic information on the 99 counties
of the state of Iowa.
}
\examples{
data(iowa)
print(iowa)
}
\concept{data}
