% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{is_contiguous}
\alias{is_contiguous}
\title{Check that a \code{redist_map} object is contiguous}
\usage{
is_contiguous(x)
}
\arguments{
\item{x}{the object}
}
\value{
\code{TRUE} if contiguous.
}
\description{
Check that a \code{redist_map} object is contiguous
}
\concept{prepare}
