% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splits.R
\name{is_county_split}
\alias{is_county_split}
\title{Identify which counties are split by a plan}
\usage{
is_county_split(plan, counties)
}
\arguments{
\item{plan}{A vector of precinct/unit assignments}

\item{counties}{A vector of county names or county ids.}
}
\value{
A logical vector which is \code{TRUE} for precincts belonging to
  counties which are split
}
\description{
Identify which counties are split by a plan
}
\concept{analyze}
