% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumpart.R
\name{is_last}
\alias{is_last}
\title{check if last edge}
\usage{
is_last(i, v, edges)
}
\arguments{
\item{i}{integer, current frontier}

\item{v}{integer, vertex to search for}

\item{edges}{edgelist matrix}
}
\value{
bool
}
\description{
check if last edge
}
