% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{last_plan}
\alias{last_plan}
\title{Extract the last plan from a set of plans}
\usage{
last_plan(plans)
}
\arguments{
\item{plans}{A \code{\link{redist_plans}} object}
}
\value{
An integer vector containing the final plan assignment.
}
\description{
Extract the last plan from a set of plans
}
\concept{analyze}
