% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_splitting.R
\name{manually_split_a_multidistrict}
\alias{manually_split_a_multidistrict}
\title{Manually Split A Multidistrict within a Plan}
\usage{
manually_split_a_multidistrict(
  map,
  counties,
  partial_plan_labels,
  region_sizes,
  region_id_to_split,
  split_dval_min = 1,
  split_dval_max = NULL,
  split_district_only,
  k_param = 1,
  verbose = F
)
}
\arguments{
\item{map}{A \code{\link[=redist_map]{redist_map()}} object.}

\item{counties}{A vector containing county (or other administrative or
geographic unit) labels for each unit, which may be integers ranging from 1
to the number of counties, or a factor or character vector.  If provided,
the algorithm will only generate spanning trees where removing an edge splits
at most one county.}

\item{partial_plan_labels}{A vector or matrix of plan id labels. It should be
zero indexed. If its a matrix then only the first column is used.}

\item{region_sizes}{A vector or matrix of region sizes. If its a matrix then
only the first column is used.}

\item{region_id_to_split}{The id of the region to split}

\item{verbose}{Whether or not to print the number of iterations until a
successful split is made and print the split plan}
}
\value{
A list with the following elements
\itemize{
\item \code{num_attempts} - The number of attempts it took to draw the tree.
\item \code{uncut_tree} - The spanning tree drawn stored as a 0-indexed directed
adjacency list.
\item \code{region_id_that_was_split} - The label of the region that was split.
\item \code{region_id_that_was_split} - The label of the region that was split.
\item \code{region_sizes} - 1-indexed region sizes.
\item \code{partial_plan_labels} - 0-indexed region labels of the split partial
plan.
\item \code{region_pops} - 1-indexed region populations.
\item \code{num_regions} - The number of regions in the split plan.
\item \code{num_districts} - The number of districts in the split plan.
\item \code{uncut_tree} - The uncut spanning tree drawn stored as a 0-indexed directed
adjacency list.
\item \code{uncut_tree_root} - The root of the uncut spanning tree drawn.
\item \code{cut_tree} - The cut spanning tree drawn stored as a 0-indexed directed
adjacency list.
\item \code{pop_below} - The population below each vertex in \code{uncut_tree} ie
the population induced by removing the edge terminating in that vertex
\item \code{uncut_tree_vertex_parents}: The parents of each zero-indexed vertex in
the tree. A value of -1 means that vertex is the root and -2 means that
vertex is not in the tree.
\item \code{new_region1_id} - The label of the first of the two new split
regions.
\item \code{new_region1_tree_root} - The root of the cut tree associated with
the first of the two new split regions.
\item \code{new_region1_size} - The size of the first of the two new split
regions.
\item \code{new_region1_pop} - The population of the first of the two new split
regions.
\item \code{new_region2_id} - The label of the second of the two new split
regions.
\item \code{new_region2_tree_root} - The root of the cut tree associated with
the second of the two new split regions.
\item \code{new_region2_size} - The size of the second of the two new split
regions.
\item \code{new_region2_pop} - The population of the second of the two new split
regions.
\item \code{cut_edge} - The two zero-indexed vertices making up the edge that
was cut from \code{uncut_tree}.
}
}
\description{
Splits a multidistrict in a plan according to the procedure in ???.
}
