% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{number_by}
\alias{number_by}
\title{Renumber districts to match a quantity of interest}
\usage{
number_by(data, x, desc = F)
}
\arguments{
\item{data}{a \code{redist_plans} object}

\item{x}{\code{\link[dplyr:dplyr_data_masking]{<data-masking>}} the quantity of interest.}

\item{desc}{\code{TRUE} if district should be sorted in descending order.}
}
\value{
a modified \code{redist_plans} object. New district numbers will be
stored as an ordered factor variable in the \code{district} column. The
district numbers in the plan matrix will match the levels of this factor.
}
\description{
District numbers in simulated plans are by and large random.  This
function will renumber the districts across all simulated plans in order
of a provided quantity of interest.
}
\concept{analyze}
