% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parity.R
\name{optimal_transfer}
\alias{optimal_transfer}
\title{Calculates Sparse Population Moves to Reduce Population Deviation to 1}
\usage{
optimal_transfer(map, plan, county_splits = FALSE, no_transfer = data.frame())
}
\arguments{
\item{map}{a \link{redist_map}}

\item{plan}{an integer vector containing the plan to be balanced.}

\item{county_splits}{a boolean value that indicates whether to recommend a
set of moves in which no additional counties are split (in testing)}

\item{no_transfer}{a dataframe consisting of a from, to, and pop column that
lists the districts between which no transfer should occur}
}
\value{
a dataframe consisting of a from, to, and pop column that describes
how much population should be moved from/to each district
}
\description{
This function computes a minimal set of population moves (e.g., 5 people from
district 1 to district 3) to balance the population between districts such
that the deviation is at most 1 between districts.
}
\examples{
data(iowa)
optimal_transfer(redist_map(iowa, existing_plan = cd_2010), iowa$cd_2010)

}
\concept{analyze}
