% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{perform_a_valid_multidistrict_split}
\alias{perform_a_valid_multidistrict_split}
\title{Split a multidistrict into two regions}
\usage{
perform_a_valid_multidistrict_split(
  adj_list,
  counties,
  pop,
  ndists,
  num_regions,
  num_districts,
  region_id_to_split,
  target,
  lower,
  upper,
  region_ids,
  region_sizes,
  split_dval_min,
  split_dval_max,
  split_district_only,
  verbose = FALSE,
  k_param = 1L
)
}
\arguments{
\item{adj_list}{A 0-indexed adjacency list representing the undirected graph
which represents the underlying map the plans are to be drawn on}

\item{counties}{Vector of county labels of each vertex in \code{g}}

\item{pop}{A vector of the population associated with each vertex in \code{g}}

\item{ndists}{The number of districts the final plans will have}

\item{target}{Ideal population of a valid district. This is what deviance is calculated
relative to}

\item{lower}{Acceptable lower bounds on a valid district's population}

\item{upper}{Acceptable upper bounds on a valid district's population}

\item{k_param}{The k parameter from the SMC algorithm, you choose among the top k_param edges}
}
\description{
Splits a multidistrict into two new regions within population bounds
}
\details{
Splits a multidistrict into two new valid regions by drawing spanning
trees uniformly at random and attempting to find an edge to cut until
a successful cut is made.
}
