% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{pl}
\alias{pl}
\title{Access the Current `redist_plans()` Object}
\usage{
pl()
}
\value{
A `redist_plans` object, or `NULL` if not called from inside a
  `dplyr` function.
}
\description{
Useful inside piped expressions and `dplyr` functions.
}
\examples{
pl()

}
