% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_compare.R
\name{plot.redist_classified}
\alias{plot.redist_classified}
\title{Plot a plan classification}
\usage{
\method{plot}{redist_classified}(x, plans, shp, type = "fill", which = NULL, ...)
}
\arguments{
\item{x}{a \code{redist_classified} object, the output of \code{\link[=classify_plans]{classify_plans()}}.}

\item{plans}{a \link{redist_plans} object.}

\item{shp}{a shapefile or \link{redist_map} object.}

\item{type}{either \code{"line"} or \code{"fill"}. Passed on to \code{\link[=compare_plans]{compare_plans()}} as
\code{plot}.}

\item{which}{indices of the splits to plot. Defaults to all}

\item{...}{passed on to \code{\link[=compare_plans]{compare_plans()}}}
}
\value{
ggplot comparison plot
}
\description{
Plot a plan classification
}
\concept{analyze}
