% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_constr.R
\name{plot.redist_constr}
\alias{plot.redist_constr}
\title{Visualize constraints}
\usage{
\method{plot}{redist_constr}(x, y, type = "group", xlim = c(0, 1), ...)
}
\arguments{
\item{x}{A [redist_constr] object.}

\item{y}{Ignored.}

\item{type}{What type of constraint to visualize. Currently supports only
`"group"`, for visualizing constraint strength by group share.}

\item{xlim}{Range of group shares to visualize.}

\item{\dots}{additional arguments (ignored)}
}
\value{
A ggplot object
}
\description{
Visualize constraints
}
