% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_map.R
\name{plot.redist_map}
\alias{plot.redist_map}
\title{Plot a \code{redist_map}}
\usage{
\method{plot}{redist_map}(x, y, ...)
}
\arguments{
\item{x}{the \code{redist_map} object}

\item{y}{\code{\link[dplyr]{<data-masking>}} the optional value to use to
color the units. If absent, \code{\link{redist.map}} will be called;
otherwise \code{\link{redist.choropleth}} will be called.}

\item{...}{passed on to underlying functions}
}
\description{
Plot a \code{redist_map}
}
\examples{
d = redist_map(fl25, n_distr=3, pop_tol=0.05)
plot(d)
plot(d, adjacency=F)
plot(d, BlackPop/pop)

}
