% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_plans.R
\name{plot.redist_plans}
\alias{plot.redist_plans}
\title{Summary plots for \code{redist_plans}}
\usage{
\method{plot}{redist_plans}(x, y, ...)
}
\arguments{
\item{x}{the \code{redist_plans} object.}

\item{y}{ignored}

\item{...}{ignored}
}
\description{
Summary plots for \code{redist_plans}
}
