% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_plans.R
\name{plot_distr_qtys}
\alias{plot_distr_qtys}
\title{Plot sorted box and whisker plots}
\usage{
plot_distr_qtys(x, qty, desc = F, ...)
}
\arguments{
\item{x}{the \code{redist_plans} object.}

\item{qty}{\code{\link[dplyr:dplyr_data_masking]{<data-masking>}} the
quantity of interest.}

\item{desc}{set \code{TRUE} to sort districts in descending order of \code{qty}}

\item{...}{passed on to \code{\link[ggplot2]{geom_boxplot}}}
}
\description{
Plots a boxplot of a quantity of interest across districts, with districts
sorted by this quantity. Adds reference points for each reference plan, if
applicable.
}
\concept{visualize}
