% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_plans.R
\name{plot_hist}
\alias{plot_hist}
\alias{hist.redist_plans}
\title{Plot a histogram of a summary statistic}
\usage{
plot_hist(x, qty, bins = ceiling(sqrt(nrow(x)) * 1.5), ...)

\method{hist}{redist_plans}(x, qty, ...)
}
\arguments{
\item{x}{the \code{redist_plans} object.}

\item{qty}{\code{\link[dplyr:dplyr_data_masking]{<data-masking>}} the statistic.}

\item{bins}{the number of bins to use in the histogram}

\item{...}{passed on to \code{\link[ggplot2]{geom_histogram}}}
}
\description{
Plots a histogram of a statistic of a \code{\link{redist_plans}} object,
with a reference line for each reference plan, if applicable.
}
\concept{plot}
