% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_plans.R
\name{plot_plan}
\alias{plot_plan}
\title{Plot a district assignment}
\usage{
plot_plan(x, draw, geom)
}
\arguments{
\item{x}{a \code{redist_plans} object.}

\item{draw}{the plan to plot. Will match the \code{draw} column of \code{x}.}

\item{geom}{the geometry to use}
}
\description{
Plot a district assignment
}
\concept{visualize}
