% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_subgraph.R
\name{preproc.adj}
\alias{preproc.adj}
\title{Preprocess adjacency arguments}
\usage{
preproc.adj(shp, adj)
}
\arguments{
\item{shp}{sf or SpatialPolygonsDataFrame}

\item{adj}{adjacency list for shp}
}
\value{
adjacency list
}
\description{
Preprocess adjacency arguments
}
\examples{
\dontrun{
data("fl25")
preproc.adj(fl25,NULL)
}
}
