% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_subgraph.R
\name{preproc.shp}
\alias{preproc.shp}
\title{Preprocess shapefile arguments}
\usage{
preproc.shp(shp)
}
\arguments{
\item{shp}{sf or SpatialPolygonsDataFrame}
}
\value{
shp as sf object
}
\description{
Preprocess shapefile arguments
}
\examples{
\dontrun{
data("fl25")
preproc.shp(fl25)
}
}
