% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.redist}
\alias{print.redist}
\title{Print legacy redist objects}
\usage{
\method{print}{redist}(x, ...)
}
\arguments{
\item{x}{redist object}

\item{\dots}{additional arguments}
}
\value{
prints to console
}
\description{
Print legacy redist objects
}
