% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_compare.R
\name{print.redist_classified}
\alias{print.redist_classified}
\title{Print redist_classified objects}
\usage{
\method{print}{redist_classified}(x, ...)
}
\arguments{
\item{x}{redist_classified object}

\item{\dots}{additional arguments}
}
\value{
prints to console
}
\description{
Print redist_classified objects
}
