% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_constr.R
\name{print.redist_constr}
\alias{print.redist_constr}
\title{Generic to print redist_constr}
\usage{
\method{print}{redist_constr}(x, header = TRUE, details = TRUE, ...)
}
\arguments{
\item{x}{redist_constr}

\item{header}{if FALSE, then suppress introduction / header line}

\item{details}{if FALSE, then suppress the details of each constraint}

\item{\dots}{additional arguments}
}
\value{
Invisibly returns the input \code{redist_constr}
}
\description{
Generic to print redist_constr
}
