% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_map.R
\name{print.redist_map}
\alias{print.redist_map}
\title{Generic to print redist_map}
\usage{
\method{print}{redist_map}(x, ...)
}
\arguments{
\item{x}{redist_map}

\item{\dots}{additional argumentss}
}
\value{
Prints to console and returns input redist_map
}
\description{
Generic to print redist_map
}
