% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_plans.R
\name{print.redist_plans}
\alias{print.redist_plans}
\title{Print method for \code{redist_plans}}
\usage{
\method{print}{redist_plans}(x, ...)
}
\arguments{
\item{x}{a [redist_plans] object}

\item{\dots}{additional arguments (ignored)}
}
\value{
The original object, invisibly.
}
\description{
Print method for \code{redist_plans}
}
