% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_plans.R
\name{rbind.redist_plans}
\alias{rbind.redist_plans}
\title{Combine multiple sets of redistricting plans}
\usage{
\method{rbind}{redist_plans}(..., deparse.level = 1)
}
\arguments{
\item{...}{The \code{\link{redist_plans}} objects to combine.  If named arguments are
provided, the names will be used in the \code{chain} column; otherwise, numbers
will be used for the \code{chain} column.}

\item{deparse.level}{Ignored.}
}
\value{
A new \code{\link{redist_plans}} object.
}
\description{
Only works when all the sets are compatible---generated from the same map,
with the same number of districts.  Sets of plans will be indexed by the
\code{chain} column.
}
\concept{analyze}
