% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_dist.R
\name{re_dist}
\alias{re_dist}
\title{Reimplementation of dist()}
\usage{
re_dist(x, method, diag = FALSE, upper = FALSE, p = 2)
}
\description{
This function reimplements the dist() function in base R
as accurately as possible.
}
