% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumpart.R
\name{redist.calc.frontier.size}
\alias{redist.calc.frontier.size}
\title{Calculate Frontier Size}
\usage{
redist.calc.frontier.size(ordered_path)
}
\arguments{
\item{ordered_path}{Path to ordered edge file created by
`redist.prep.enumpart()`.}
}
\value{
A list with `max`, `average`, `average_sq`, and `sequence`.
}
\description{
**Deprecated**: Use [redist_enumpart_frontier()] instead.
}
\references{
Fifield, B., Imai, K., Kawahara, J., & Kenny, C. T. (2020).
The essential role of empirical validation in legislative redistricting
simulation. *Statistics and Public Policy*, 7(1), 52-68.
}
\concept{enumerate}
\keyword{internal}
