% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcstats.R
\name{redist.calcstats}
\alias{redist.calcstats}
\title{Calculate standard redistricting diagnostic statistics}
\usage{
redist.calcstats(algout, group1vote, group2vote, margin, swing, npoints, stats)
}
\arguments{
\item{algout}{An object of class "redist" or a matrix of congressional district
assignments where columns are different redistricting plans.}

\item{group1vote}{Vote counts for group 1 (for instance, Democrats)}

\item{group2vote}{Vote counts for group 2 (for instance, Republicans)}

\item{margin_mc}{The margin (50% - margin, 50% + margin) for counting "close" plans
for the "mc" stat.}

\item{margin_ps}{The margin (50% - swing, 50% + swing) over which to calculate the
partisan symmetry statistic.}

\item{npoints}{The number of points along which to estimate the integral for the
partisan symmetry statistic.}

\item{stats}{The stats to calculate. "ec" refers to electoral competitiveness,
"mc" is a count of plans within a particular vote margin, "ps_simple" is a
simple measure of partisan symmetry that counts the expected number of seats
given a party vote share minus the expected number of seats that party would receive
under the other party's vote share, "eg" is the efficiency gap, and "ps_full" is a
full implementation of the partisan symmetry diagnostic that measures the area
between the empirical seats-votes curve and a null, balanced seats-votes curve.}

\item{nc}{Number of cores to parallelize calculation over. Default is detectCores() - 1.}
}
\value{
\code{redist.calcstats} returns a list of estimated statistics for each plan.
}
\description{
\code{redist.calcstats} can calculate various statistics used to
summarize ensembles of simulated redistricting plans.
}
