% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{redist.coarsen.adjacency}
\alias{redist.coarsen.adjacency}
\title{Coarsen Adjacency List}
\usage{
redist.coarsen.adjacency(adj, groups)
}
\arguments{
\item{adj}{A zero-indexed adjacency list. Required.}

\item{groups}{integer vector of elements of adjacency to group}
}
\value{
adjacency list coarsened
}
\description{
Coarsen Adjacency List
}
\concept{prepare}
