% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitiveness.R, R/tidy.R
\name{redist.competitiveness}
\alias{redist.competitiveness}
\alias{competitiveness}
\title{Compute Competitiveness}
\usage{
redist.competitiveness(plans, rvote, dvote)

competitiveness(map, rvote, dvote, .data = get0(".", parent.frame()))
}
\arguments{
\item{plans}{A numeric vector (if only one map) or matrix with one row
for each precinct and one column for each map. Required.}

\item{rvote}{A numeric vector with the Republican vote for each precinct.}

\item{dvote}{A numeric vector with the Democratic vote for each precinct.}

\item{map}{a \code{\link{redist_map}} object}

\item{.data}{a \code{\link{redist_plans}} object}
}
\value{
Numeric vector with competitiveness scores
}
\description{
Currently only implements the competitiveness function in equation (5)
of Cho & Liu 2016.
}
\examples{
\dontrun{
data(fl25)
data(fl25_enum)

plans_05 <- fl25_enum$plans[, fl25_enum$pop_dev <= 0.05]
comp <- redist.competitiveness(plans_05, fl25$mccain, fl25$obama)
}
}
\concept{analyze}
