% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findtarget.R
\name{redist.constraint.helper}
\alias{redist.constraint.helper}
\title{Create Constraints for SMC}
\usage{
redist.constraint.helper(
  constraints = "vra",
  tgt_min = 0.55,
  group_pop,
  total_pop,
  ndists,
  nmmd,
  strength_vra = 2500,
  pow_vra = 1.5
)
}
\arguments{
\item{constraints}{Vector of constraints to include. Currently only 'vra' implemented.}

\item{tgt_min}{Defaults to 0.55. If 'vra' included, the minority percent to encourage in each district.}

\item{group_pop}{A vector of populations for some subgroup of interest.}

\item{total_pop}{A vector containing the populations of each geographic unit.}

\item{ndists}{The total number of districts.}

\item{nmmd}{The number of majority minority districts to target for 'vra' constraint}

\item{strength_vra}{The strength of the 'vra' constraint. Defaults to 2500.}

\item{pow_vra}{The exponent for the 'vra' constraint. Defaults to 1.5.}
}
\value{
list of lists for each constraint selected
}
\description{
Create Constraints for SMC
}
\concept{prepare}
