% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counties.R
\name{redist.county.id}
\alias{redist.county.id}
\title{Create County IDs}
\usage{
redist.county.id(counties)
}
\arguments{
\item{counties}{vector of counties, required.}
}
\value{
A vector with an ID that corresponds from 1:n counties
}
\description{
Create County IDs
}
\examples{
set.seed(2)
counties <- sample(c(rep('a', 20), rep('b', 5)))
redist.county.id(counties)

}
\concept{prepare}
