% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_viz.R
\name{redist.group_viz}
\alias{redist.group_viz}
\title{Visualize Plans and Their Treatment of Groups}
\usage{
redist.group_viz(prep, browser = TRUE)
}
\arguments{
\item{prep}{The output of the \code{\link{redist.prep_group_viz}} function,
which handles all the computations and preparatory work.}

\item{browser}{if \code{TRUE} (the default), launch the Shiny app in the
default browser. Otherwise, just start the Shiny server.}
}
\description{
Launches a \code{shiny} app which allows a user to explore a set of plans.
}
\details{
The app plots plans as points on a two-dimensional plane, based on
multidimensional scaling applied to a variation of information distance
matrix (see \code{\link{redist.distances}}).  Users can click any point on the
plane to see the corresponding plan.  Points are colored by their share of
the group population (where the group can be minority voters, Democratic
voters, or some other group).  The app also clusters plans according to their
similarity, and plots a representative plan for each cluster, allowing users
to quickly get a sense of the high-level structure of the space of possible
redistricting plans.
}
\examples{
\dontrun{
data("fl25")
data(algdat.p10)
# Sample
sampled_plans = redist.smc(algdat.p10$adjlist, fl25$TotPop,
                           nsims=200, ndists=3, popcons=0.1)

# Prep
viz_prep = redist.prep_group_viz(sampled_plans, fl25, fl25$VAP, fl25$BlackVAP)

# Run the app!
redist.group_viz(viz_prep)
}
}
