% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vraplots.R
\name{redist.majmin.plot}
\alias{redist.majmin.plot}
\title{Majority Minority Plots}
\usage{
redist.majmin.plot(grouppercent, type = "hist", title = "")
}
\arguments{
\item{grouppercent}{output from redist.group.percent}

\item{type}{string in 'hist', 'toptwo', or 'box'}

\item{title}{ggplot title}
}
\value{
ggplot
}
\description{
Majority Minority Plots
}
