% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splits.R
\name{redist.muni.splits}
\alias{redist.muni.splits}
\title{Counts the Number of Municipalities Split Between Districts}
\usage{
redist.muni.splits(plans, munis)
}
\arguments{
\item{plans}{A numeric vector (if only one map) or matrix with one row
for each precinct and one column for each map. Required.}

\item{munis}{A vector of municipality names or ids.}
}
\value{
integer matrix where each district is a
}
\description{
Counts the total number of municpalities that are split.
Municipalities in this interpretation do not need to cover the entire state, which
differs from counties.
}
\examples{
data(iowa)
ia <- redist_map(iowa, existing_plan = cd_2010, total_pop = pop, pop_tol = 0.01)
plans <- redist_smc(ia, 50, silent = TRUE)
ia$region[1:10] <- NA
splits <- redist.muni.splits(plans, ia$region)
}
\concept{analyze}
