% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parity.R
\name{redist.parity}
\alias{redist.parity}
\title{Calculates Maximum Deviation from Population Parity}
\usage{
redist.parity(district_membership, population, ncores = 1)
}
\arguments{
\item{district_membership}{A matrix with one row 
for each precinct and one column for each map. Required.}

\item{population}{A numeric vector with the population for every precinct.}

\item{ncores}{Number of cores to use for parallel computing. Default is 1.}
}
\value{
One deviation metric for one plan. If a matrix of plans is provided,
 returns a vector of the deviation of each plan.
}
\description{
Computes the deviation from population parity from a plan.
Higher values indicate that (at least) a single district in the map deviates 
from population parity. See Details.
}
\details{
With a map with \code{pop} representing the populations of each district,
 the deviation from population parity is given as \code{max(abs(pop - parity) / parity)} 
 where \code{parity = sum(pop)/length(pop)} is the population size for the 
 average district.
 Therefore, the metric can be thought of as the maximum percent deviation from 
 equal population. For example, a value of 0.03 in this metric indicates that
 all districts are within 3 percent of population parity.
}
\examples{
data("fl25")
data("algdat.p10")
# Deviation from parity in 5 plans
redist.parity(district_membership = algdat.p10$cdmat[, 1:5], 
              population = algdat.p10$precinct.data$pop)

}
