% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{redist.plot.adj}
\alias{redist.plot.adj}
\title{Creates a Graph Overlay}
\usage{
redist.plot.adj(
  shp = NULL,
  adj = NULL,
  plan = NULL,
  centroids = TRUE,
  drop = FALSE,
  title = ""
)
}
\arguments{
\item{shp}{A SpatialPolygonsDataFrame or sf object. Required.}

\item{adj}{A zero-indexed adjacency list. Created with redist.adjacency
if not supplied. Default is NULL.}

\item{plan}{A numeric vector with one entry for each precinct in shp.
Used to remove edges that cross boundaries. Default is \code{NULL}.  Optional.}

\item{centroids}{A logical indicating if centroids should be plotted. Default is \code{TRUE}.}

\item{drop}{A logical indicating if edges that cross districts should be dropped. Default is \code{FALSE}.}

\item{title}{A string title of plot. Defaults to empty string. Optional.}
}
\value{
ggplot map
}
\description{
Creates a Graph Overlay
}
\examples{
\dontrun{
data(iowa)
redist.plot.adj(shp = iowa, plan = iowa$cd_2010)
}

}
\concept{plot}
