% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_plot.R
\name{redist.plot.cores}
\alias{redist.plot.cores}
\title{Plot Cores}
\usage{
redist.plot.cores(shp, plan, core)
}
\arguments{
\item{shp}{A SpatialPolygonsDataFrame or sf object. Required.}

\item{plan}{A numeric vector with one entry for each precinct in shp.
Used to color the districts. Required.}

\item{core}{Required. integer vector produced by \code{redist.identify.cores()}.}
}
\value{
ggplot
}
\description{
Plot Cores
}
\concept{plot}
