% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{redist.plot.map}
\alias{redist.plot.map}
\title{Plot a Map}
\usage{
redist.plot.map(
  shp,
  adj,
  plan = NULL,
  fill = NULL,
  fill_label = "",
  boundaries = TRUE,
  title = ""
)
}
\arguments{
\item{shp}{A SpatialPolygonsDataFrame, sf object, or redist_map. Required.}

\item{adj}{A zero-indexed adjacency list. Created with redist.adjacency
if not supplied and needed for coloring. Default is NULL.}

\item{plan}{A numeric vector with one entry for each precinct in shp.
Used to color the districts. Default is \code{NULL}.  Optional.}

\item{fill}{A numeric/integer vector with values to color the plot with. Optional.}

\item{fill_label}{A string title of plot. Defaults to empty string.}

\item{boundaries}{A logical indicating if precinct boundaries should be plotted.}

\item{title}{A string title of plot. Defaults to empty string. Optional.}
}
\value{
ggplot map
}
\description{
Create a ggplot map. It fills by plan or argument fill. If both are supplied,
plan is used as the color and fill as the alpha parameter.
}
\examples{
\dontrun{
data(iowa)
redist.plot.map(shp = iowa, plan = iowa$cd_2010)

iowa \%>\% redist_map(existing_plan = cd_2010) \%>\% redist.plot.map(shp = ., plan = get_existing(.))
}

}
\concept{plot}
