% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penaltyviz.R
\name{redist.plot.penalty}
\alias{redist.plot.penalty}
\title{Visualize VRA Penalty}
\usage{
redist.plot.penalty(
  tgt_min = 0.55,
  tgt_other = 0.25,
  strength_vra = 2500,
  pow_vra = 1.5,
  limits = TRUE
)
}
\arguments{
\item{tgt_min}{double, defaults to 0.55. The minority target percent.}

\item{tgt_other}{double, defaults to 0.25. The other group target percent.}

\item{strength_vra}{double, strength of the VRA constraint.}

\item{pow_vra}{double, exponent of the VRA constraint.}

\item{limits}{Whether to limit y axis to 0,500. Default is TRUE for 
comparability across values.}
}
\value{
ggplot
}
\description{
Plots the shape of the VRA Gibbs penalty
}
\examples{
redist.plot.penalty()

}
