% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{redist.plot.plan}
\alias{redist.plot.plan}
\title{Plot a Map}
\usage{
redist.plot.plan(shp, adj, plan = NULL, boundaries = TRUE, title = "")
}
\arguments{
\item{shp}{A SpatialPolygonsDataFrame or sf object. Required.}

\item{adj}{A zero-indexed adjacency list. Created with redist.adjacency
if not supplied. Default is NULL.}

\item{plan}{A numeric vector with one entry for each precinct in shp.
Used to color the districts. Default is \code{NULL}.  Optional.}

\item{boundaries}{A logical indicating if precinct boundaries should be plotted.}

\item{title}{A string title of plot. Defaults to empty string. Optional.}
}
\value{
ggplot map
}
\description{
Plot a Map
}
\examples{
\dontrun{
data(iowa)
redist.plot.plan(shp = iowa, plan = iowa$cd_2010)

iowa \%>\% redist_map(existing_plan = cd_2010) \%>\% redist.plot.plan(shp = ., plan = get_existing(.))
}

}
\concept{plot}
