% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_plans.R
\name{redist.plot.plans}
\alias{redist.plot.plans}
\title{Plot a district assignment}
\usage{
redist.plot.plans(x, draws, geom)
}
\arguments{
\item{x}{a \code{redist_plans} object.}

\item{draws}{the plan(s) to plot. Will match the \code{draw} column of \code{x}.}

\item{geom}{the geometry to use}
}
\description{
Plot a district assignment
}
\concept{plot}
