% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_varinfo.R
\name{redist.plot.varinfo}
\alias{redist.plot.varinfo}
\title{Static Variation of Information Plot}
\usage{
redist.plot.varinfo(plans, group_pop, total_pop, shp)
}
\arguments{
\item{plans}{matrix of district assignments}

\item{group_pop}{Required Population of subgroup being studied in each precinct.}

\item{total_pop}{Required. Population of each precinct.}

\item{shp}{sf dataframe}
}
\value{
patchworked ggplot
}
\description{
Static Variation of Information Plot
}
\concept{plot}
