% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcstats.R
\name{redist.polsbypopper}
\alias{redist.polsbypopper}
\title{Polsby-Popper calculation for MCMC}
\usage{
redist.polsbypopper(algout, adj_list, areas_vec, borderlength_mat, pop_vec,
  discrete = FALSE)
}
\description{
Polsby-Popper calculation for MCMC
}
