% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumpart.R
\name{redist.prep.enumpart}
\alias{redist.prep.enumpart}
\title{Prepare enumpart edge ordering}
\usage{
redist.prep.enumpart(
  adj,
  ordered_path,
  weight_path = NULL,
  total_pop = NULL,
  unordered_path
)
}
\arguments{
\item{adj}{Zero-indexed adjacency list.}

\item{ordered_path}{Valid path to output the ordered adjacency map to.}

\item{weight_path}{A path (not including `".dat"`) to store vertex weights.
Only supply with `total_pop`.}

\item{total_pop}{The vector of precinct populations. Only supply with
`weight_path`.}

\item{unordered_path}{Deprecated, ignored.}
}
\value{
0 on success.
}
\description{
**Deprecated**: Use [redist_enumpart()] instead, which handles preparation
automatically.
}
\references{
Fifield, B., Imai, K., Kawahara, J., & Kenny, C. T. (2020).
The essential role of empirical validation in legislative redistricting
simulation. *Statistics and Public Policy*, 7(1), 52-68.
}
\concept{enumerate}
\keyword{internal}
