% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness.R
\name{redist.prep.polsbypopper}
\alias{redist.prep.polsbypopper}
\title{Prep Polsby Popper Perimeter Dataframe}
\usage{
redist.prep.polsbypopper(shp, adj, perim_path, ncores = 1)
}
\arguments{
\item{shp}{A SpatialPolygonsDataFrame or sf object. Required unless "EdgesRemoved"
and "logSpanningTree" with adjacency provided.}

\item{adj}{Required, A zero-indexed adjacency list}

\item{perim_path}{A path to save an Rds}

\item{ncores}{the number of cores to parallelize over}
}
\value{
A perimeter dataframe
}
\description{
Prep Polsby Popper Perimeter Dataframe
}
\examples{
\dontrun{
data(fl25)
data("algdat.p10")
adj <- redist.adjacency(fl25)
plans <- algdat.p10$cdmat
perim_df <- redist.prep.polsbypopper(shp = fl25, adj = adj)
}
}
