% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_viz.R
\name{redist.prep_group_viz}
\alias{redist.prep_group_viz}
\title{Prepare to Visualize Plans}
\usage{
redist.prep_group_viz(
  result,
  geometry,
  full_pop,
  group_pop,
  current = NULL,
  renumber = TRUE,
  by = "people",
  k = NULL,
  simp_geom = TRUE
)
}
\arguments{
\item{result}{a \code{redist} object produced by \code{\link{redist.smc}}
or \code{\link{redist.mcmc}}, or a matrix of district assignments.}

\item{geometry}{an \code{sf} data frame containing shapefile information for
the current map.}

\item{full_pop}{a numeric vector with the population of the group for every precinct.}

\item{group_pop}{a numeric vector with the population of every precinct.}

\item{current}{a numeric vector containing the current plan, which will be plotted separately.}

\item{renumber}{whether to renumber districts in ascending order by their group share, so
district 1 has the smallest fraction of the group, district 2 has the second-smallest,
and so on.}

\item{by}{the way to compute the variation of information distance, by \code{"people"},
the recommended default, by \code{"group"} or \code{"minority"}, or by
\code{"land"}, which treats each precinct equally.}

\item{k}{the number of clusters to group the plans into. If left unspecified,
this will be determined automatically using the silhouette method.}

\item{simp_geom}{if \code{TRUE}, the geometry will be slightly simplified to
allow for faster plotting.}
}
\value{
A list of class \code{redist_prep_viz} containing the preprocessed
  inputs and calculations.
}
\description{
This helper function performs all necessary computations and preparation to
run the Shiny app in \code{\link{redist.group_viz}}
}
\examples{
\dontrun{
data("fl25")
data(algdat.p10)
# Sample
sampled_basic = redist.smc(algdat.p10$adjlist, fl25$TotPop,
                           nsims=200, ndists=3, popcons=0.1)

# Prep
viz_prep = redist.prep_group_viz(sampled_plans, fl25$geometry, fl25$VAP, fl25$BlackVAP)

# Run the app!
redist.group_viz(viz_prep)
}
}
