% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{redist.reduce.adjacency}
\alias{redist.reduce.adjacency}
\title{Reduce Adjacency List}
\usage{
redist.reduce.adjacency(adjacency, keep_rows)
}
\arguments{
\item{adjacency}{A zero-indexed adjacency list. Required.}

\item{keep_rows}{row numbers of precincts to keep}
}
\value{
zero indexed adjacency list with max value length(keep_rows) - 1
}
\description{
Tool to help reduce adjacency lists for analyzing subsets of maps.
}
\examples{
\dontrun{
data("algdat.p10")
redist.reduce.adjacency(algdat.p10$adjlist, c(2, 3, 4, 6, 21))
}
}
