% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate.R
\name{redist.samplepart}
\alias{redist.samplepart}
\title{Sample partitions using spanning trees}
\usage{
redist.samplepart(adjobj, ndists, popcons, popvec, contiguitymap)
}
\arguments{
\item{adjobj}{An adjacency list, matrix, or object of class
\code{SpatialPolygonsDataFrame}.}

\item{ndists}{The desired number of congressional districts}

\item{popcons}{The strength of the hard population constraint.
\code{popcons} = 0.05 means that any sample dtree will be within
5% of population parity. The default is
\code{NULL}.}

\item{popvec}{A vector of geographic unit populations. The default is
\code{NULL}. Must be specified if popcons is not null.}

\item{contiguitymap}{Use queens or rooks distance criteria for generating an
adjacency list from a "SpatialPolygonsDataFrame" data type.
Default is "rooks".}
}
\value{
\code{redist.samplepart} returns a list where the first entry is the
randomly sampled redistricting plan, and the second entry is the number of
possible redistricting plans from the implied spanning tree.
}
\description{
\code{redist.samplepart} uses a spanning tree method to randomly sample
redistricting plans.
}
