% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{redist.smc_is_ci}
\alias{redist.smc_is_ci}
\title{(Deprecated) Confidence Intervals for Importance Sampling Estimates}
\usage{
redist.smc_is_ci(x, wgt, conf = 0.99)
}
\arguments{
\item{x}{A numeric vector containing the quantity of interest}

\item{wgt}{A numeric vector containing the nonnegative importance weights.
Will be normalized automatically.}

\item{conf}{The confidence level for the interval.}
}
\value{
A two-element vector of the form `[lower, upper]` containing
the importance sampling confidence interval.
}
\description{
Builds a confidence interval for a quantity of interest,
given importance sampling weights.
}
\concept{post}
