% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splits.R, R/tidy.R
\name{redist.splits}
\alias{redist.splits}
\alias{county_splits}
\title{Count County Splits}
\usage{
redist.splits(plans, district_membership, counties)

county_splits(map, counties, .data = get0(".", parent.frame()))
}
\arguments{
\item{plans}{A numeric vector (if only one map) or matrix with one row
for each precinct and one column for each map. Required.}

\item{district_membership}{Deprecated, use plans. A numeric vector (if only one map) or matrix with one row
for each precinct and one column for each map. Required.}

\item{counties}{A vector of county names or county ids.}

\item{map}{a \code{\link{redist_map}} object}

\item{.data}{a \code{\link{redist_plans}} object}
}
\value{
integer vector with one number for each map
}
\description{
Count County Splits
}
\concept{analyze}
