% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist-smc.R
\name{redist.subset_bounds}
\alias{redist.subset_bounds}
\title{Calculate Population Bounds for a Subset of a Map}
\usage{
redist.subset_bounds(total_pop, subset, ndists, sub_distr, tol = 0.01)
}
\arguments{
\item{total_pop}{a numeric vector containing the population of every precinct in
the full map.}

\item{subset}{an indexing vector for the subset of the map that will be
studied or resampled.  Often this will take the form \code{current_plan
\%in\% c(1, 3, 7)}, where 1, 3, and 7 are the study districts.}

\item{ndists}{the number of districts in the overall map.}

\item{sub_distr}{the number of districts in the map subset.}

\item{tol}{the overall population tolerance that must be met.}
}
\value{
a 3-element numeric vector of the form \code{c(lower, target, upper)},
which can be used inside \code{\link{redist.smc}}.
}
\description{
Creates a 3-element vector, suitable for the precise population constraints
in \code{\link{redist.smc}}, that ensures that samples of plans built from
parts of a map will still satisfy the overall population constraint.
}
\examples{
data("fl25")
data(algdat.p10)

subset = algdat.p10$cdmat[,1] \%in\% c(1, 2)
redist.subset_bounds(fl25$TotPop, subset, 3, 2, 0.1)
}
\concept{prepare}
