% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cores.R
\name{redist.uncoarsen}
\alias{redist.uncoarsen}
\title{Uncoarsen a District Matrix}
\usage{
redist.uncoarsen(plans, group_index)
}
\arguments{
\item{plans}{A coarsened matrix of plans.}

\item{group_index}{The index used to coarsen the shape.}
}
\value{
matrix
}
\description{
After a cores analysis or other form of coarsening, sometimes you need
to be at the original geography level to be comparable. This takes in a
coarsened matrix and uncoarsens it to the original level
}
\concept{post}
