% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varinfo.R
\name{redist.varinfo.plot}
\alias{redist.varinfo.plot}
\title{Static Variation of Information Plot}
\usage{
redist.varinfo.plot(district_membership, grouppop, fullpop, shp)
}
\arguments{
\item{district_membership}{matrix of district assignments}

\item{grouppop}{Required. Population of subgroup being studied in each precinct.}

\item{fullpop}{Required. Population of each precinct.}

\item{shp}{sf dataframe}
}
\value{
patchworked ggplot
}
\description{
Static Variation of Information Plot
}
