% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weighted_adj.R
\name{redist.wted.adj}
\alias{redist.wted.adj}
\title{Create Weighted Adjacency Data}
\usage{
redist.wted.adj(map = NULL, plans = NULL)
}
\arguments{
\item{map}{redist_map}

\item{plans}{redist_plans}
}
\value{
tibble
}
\description{
Create Weighted Adjacency Data
}
\examples{
data(iowa)
shp <- redist_map(iowa, existing_plan = cd_2010, pop_tol = 0.01)
plans <- redist_smc(shp, 100)
redist.wted.adj(shp, plans = plans)
}
