% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_mew.R
\name{redist_mew}
\alias{redist_mew}
\title{Marked Edge Walk MCMC Redistricting Sampler (McWhorter and DeFord 2024)}
\usage{
redist_mew(
  map,
  nsims,
  warmup = 0L,
  thin = 1L,
  init_plan = NULL,
  compactness = 1,
  constraints = list(),
  adapt_k_thresh = 0.99,
  warmstart = FALSE,
  init_name = NULL,
  verbose = FALSE,
  silent = FALSE
)
}
\arguments{
\item{map}{A \link{redist_map} object.}

\item{nsims}{The number of samples to draw, including warmup.}

\item{warmup}{The number of warmup samples to discard.}

\item{thin}{Save every \code{thin}-th sample. Defaults to no thinning (1).}

\item{init_plan}{The initial state of the map. If not provided, will default
to the reference map of the \code{map} object, or if none exists, will
sample a random initial state using \code{\link[=redist_smc]{redist_smc()}}. You can also
request a random initial state by setting \code{init_plan="sample"}.}

\item{compactness}{Controls the compactness of the generated districts, with
higher values preferring more compact districts. Must be nonnegative.}

\item{constraints}{A \link{redist_constr} object or a list containing
information on sampling constraints. See \link{redist_constr} for more
information.}

\item{adapt_k_thresh}{The threshold value used in the heuristic to adapt the
number of marked edges. Higher values are more conservative. Set to 0.9999
or 1 if the algorithm does not appear to be sampling from the target
distribution. Must be between 0 and 1.}

\item{warmstart}{If \code{TRUE}, the algorithm will initialize the spanning
tree from the structure of \code{init_plan}. If \code{FALSE} (the
default), uses Wilson's algorithm to sample a uniform random spanning tree,
then finds marked edges that match \code{init_plan}.}

\item{init_name}{A name for the initial plan, or \code{FALSE} to not include
the initial plan in the output. Defaults to the column name of the existing
plan, or \code{"<init>"} if the initial plan is sampled.}

\item{verbose}{Whether to print out intermediate information while sampling.
Recommended.}

\item{silent}{Whether to suppress all diagnostic information.}
}
\value{
\code{redist_mew} returns a \link{redist_plans} object containing the
simulated plans.
}
\description{
\code{redist_mew} uses a Markov Chain Monte Carlo algorithm based on marked
edge walks on spanning trees (McWhorter and DeFord 2024) to generate
congressional or legislative redistricting plans according to contiguity,
population, compactness, and other custom constraints.
}
\examples{
data(fl25)
fl_map <- redist_map(fl25, ndists = 3, pop_tol = 0.1)
redist_mew(fl_map, 5000)

}
\references{
McWhorter, A., & DeFord, D. (2025). The Marked Edge Walk: A Novel MCMC
Algorithm for Sampling of Graph Partitions.
}
\concept{simulate}
