% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_plans.R
\name{redist_plans}
\alias{redist_plans}
\title{A set of redistricting plans}
\description{
A \code{redist_plans} object is essentially a data frame of summary
information on each district and each plan, along with the matrix of district
assignments and information about the simulation process used to generate the
plans.
}
\details{
The first two columns of the data frame will be \code{draw}, a factor indexing
the simulation draw, and \code{district}, an integer indexing the districts
within a plan. The data frame will therefore have \code{n_sims*n_distr} rows.
As a data frame, the usual \code{\link{dplyr}} methods will work.

Other useful methods for \code{redist_plans} objects:
\itemize{
\item \code{\link{add_reference}}
\item \code{\link{pullback}}
\item \code{\link{number_by}}
\item \code{\link{match_numbers}}
\item \code{\link{prec_assignment}}
\item \code{\link{get_plan_matrix}}
\item \code{\link{get_plan_weights}}
\item \code{\link{get_sampling_info}}
\item \code{\link{subset_sampled}}
\item \code{\link{subset_ref}}
\item \code{\link{as.matrix.redist_plans}}
\item \code{\link{plot.redist_plans}}
}
}
\concept{analyze}
