% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_smc_tidy.R
\name{redist_quantile_trunc}
\alias{redist_quantile_trunc}
\title{Helper function to truncate importance weights.}
\usage{
redist_quantile_trunc(x)
}
\arguments{
\item{x}{the weights}
}
\description{
Helper function to truncate importance weights.
}
