% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_shortburst.R
\name{scorer-arith}
\alias{scorer-arith}
\alias{*.redist_scorer}
\alias{+.redist_scorer}
\alias{-.redist_scorer}
\title{Scoring function arithmetic}
\usage{
\method{*}{redist_scorer}(x, fn2)

\method{+}{redist_scorer}(fn1, fn2)

\method{-}{redist_scorer}(fn1, fn2)
}
\arguments{
\item{x}{a numeric or a \code{redist_scorer} function, from \code{\link{scorers}}}

\item{fn2}{a \code{redist_scorer} function, from \code{\link{scorers}}}

\item{fn1}{a \code{redist_scorer} function, from \code{\link{scorers}}}
}
\value{
function of class redist_scorer
}
\description{
\code{redist_scorer} functions may be multiplied by constants and/or added
together to form linear combinations.
}
\concept{prepare}
