% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_map.R
\name{set_pop_tol}
\alias{set_pop_tol}
\title{Set the population tolerance from a \code{redist_map} object}
\usage{
set_pop_tol(map, pop_tol)
}
\arguments{
\item{map}{the \code{redist_map} object}

\item{pop_tol}{the population tolerance}
}
\value{
an updated redist_map object
}
\description{
Set the population tolerance from a \code{redist_map} object
}
\concept{prepare}
