% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skinny_flips.R
\name{skinny_flips}
\alias{skinny_flips}
\title{Dangerous but Skinny Flip}
\usage{
skinny_flips(
  adj,
  init_plan,
  total_pop,
  pop_tol,
  nsims,
  eprob,
  lambda,
  constraints
)
}
\arguments{
\item{adj}{zero indexed adjacency list}

\item{init_plan}{initial plan}

\item{total_pop}{total population}

\item{pop_tol}{maximum population deviance allowed}

\item{nsims}{number of steps to take}

\item{eprob}{edge cut probability}

\item{lambda}{number of components to swap}

\item{constraints}{constraint list}
}
\value{
matrix
}
\description{
Runs flip silently without checking inpuut quality for use within other contexts
which already check things. It returns just a matrix of plans
}
