% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solve_hungarian}
\alias{solve_hungarian}
\title{Hungarian Algorithm Solver}
\usage{
solve_hungarian(costMatrix)
}
\arguments{
\item{costMatrix}{matrix giving cost of each possible pairing - can be rectangular}
}
\value{
List with cost and parings, pairings are given as an Nx2 matrix
giving edges that are matched (1-indexed rather than 0-indexed as it will be returned to R)
}
\description{
Wrapper code from Justin Silverman, 2019
}
\details{
Solves weighted bipartite matching problems (e.g., optimal matching of people to cars
or optimal matching of students to colleges, etc...)

this is a copy/wrapper for the code developed by Cong Ma and made available
as a github repository (mcximing/hungarian-algorithm-cpp). Code was
changed to a header only file for use in other Rcpp packages.
}
