% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_splitting.R
\name{split_entire_map_into_plan}
\alias{split_entire_map_into_plan}
\title{Splits a map into \code{ndist} districts}
\usage{
split_entire_map_into_plan(
  map,
  counties,
  min_region_cut_sizes = NULL,
  max_region_cut_sizes = NULL,
  split_district_only = F,
  k_param = 1,
  verbose = F
)
}
\arguments{
\item{map}{A \code{\link[=redist_map]{redist_map()}} object.}

\item{counties}{A vector containing county (or other administrative or
geographic unit) labels for each unit, which may be integers ranging from 1
to the number of counties, or a factor or character vector.  If provided,
the algorithm will only generate maps which split up to \code{ndists-1}
counties. Even if there are fewer counties than \code{ndists - 1}, the spanning
trees will change the results of the simulations. There is no strength
parameter associated with this constraint. To adjust the number of county
splits further, or to constrain a second type of administrative split,
consider using \code{add_constr_splits()}, \code{add_constr_multisplits()}, and
\code{add_constr_total_splits()}.}

\item{verbose}{Whether to print out intermediate information while sampling.
Recommended.}
}
\value{
A list of length \code{ndists - 1} where each element of the list is
itself a list where each elements are the return values of
\code{\link{manually_split_a_multidistrict}} along with two new additional
elements of the list:
\itemize{
\item \code{cut_edges} - A list of 2-element vectors (u,v) where u,v are the zero-indexed
vertices of the edges removed from the drawn spanning trees at each step to
create the two new regions
\item \code{spanning_forest} - The spanning forest of the cut spanning trees drawn up
to that point.
}
}
\description{
Splits a map all the way into \code{ndist} districts and saves the splitting
information from each step including the pre-cut and cut spanning tree drawn
at each step that was successfully split. Very useful for debugging and
visual inspection of internal results.
}
\details{
see \code{\link{manually_split_a_multidistrict}}.
}
