% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans_helpers.R
\name{tally_var}
\alias{tally_var}
\title{Tally a variable by district}
\usage{
tally_var(map, x, .data = pl())
}
\arguments{
\item{map}{a `redist_map` object}

\item{x}{a variable to tally. Tidy-evaluated.}

\item{.data}{a `redist_plans` object or matrix of plans}
}
\value{
a vector containing the tallied values by district and plan (column-major)
}
\description{
Tally a variable by district
}
\concept{analyze}
