% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_smc_alg_helpers.R
\name{validate_counties}
\alias{validate_counties}
\title{Checks a county label input is valid and if so returns it.}
\usage{
validate_counties(map, adj_list, V, counties, counties_q = NULL)
}
\arguments{
\item{map}{A \code{redist_map}}

\item{V}{The number of vertices in the plan graph.}

\item{counties}{County thing}

\item{counties_q}{The quosure in the county label input ie
\code{counties_q <- rlang::enquo(counties)} must have been called in
the nested function.}
}
\value{
A counties label vector
}
\description{
Checks that a county label input is valid and if so returns it.
}
