% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_smc_alg_helpers.R
\name{validate_cut_sizes}
\alias{validate_cut_sizes}
\title{Crude check that cut sizes are valid}
\usage{
validate_cut_sizes(
  ndists,
  num_splits,
  min_region_cut_sizes,
  max_region_cut_sizes
)
}
\arguments{
\item{ndists}{The number of districts in the final plan.}

\item{num_splits}{The number of splits to be performed.}

\item{min_region_cut_sizes}{Vector of minimum region cut sizes
for each split.}

\item{max_region_cut_sizes}{Vector of maximum region cut sizes
for each split.}
}
\value{
A counties label vector
}
\description{
Crude check that the minimum and maximum cut sizes are valid
}
