% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_alg_helpers.R
\name{validate_initial_region_id_mat}
\alias{validate_initial_region_id_mat}
\title{Checks that a region id matrix is valid}
\usage{
validate_initial_region_id_mat(init_region_ids_mat, V, nsims, init_num_regions)
}
\arguments{
\item{init_region_ids_mat}{A V by nsims matrix of zero-indexed partial plans}

\item{V}{The number of vertices in the plan graph.}

\item{nsims}{The number of simulations being run.}

\item{@init_num_regions}{The number of regions in the partial plans stored in the
\code{init_region_ids_mat}}
}
\description{
Checks that every plan in a matrix of partial plans is valid. A partial
plans matrix with \code{init_num_regions} is valid if it is a \code{V} by \code{nsims}
matrix where for each column it only has values between
0,...,\code{init_num_regions}. Will throw an error if anything is wrong.
}
