% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_alg_helpers.R
\name{validate_sample_space_and_splitting_method}
\alias{validate_sample_space_and_splitting_method}
\title{Checks the sampling space and splitting method are valid}
\usage{
validate_sample_space_and_splitting_method(
  sampling_space,
  splitting_method,
  splitting_params,
  num_splitting_steps
)
}
\arguments{
\item{sampling_space}{The space to sample the plans on. This does not affect
the plans output by the function but the sample space used can have a large
impact on computational cost/runtime and convergence. Current spaces supported
right now are
\itemize{
\item graph_plan_space : graph partition space
\item spanning_forest_space : spanning forest space
\item linking_edge_space : linking edge forest space
}}

\item{splitting_params}{A list of parameters related to splitting the plans.
Options include
\itemize{
\item \code{splitting_schedule} What rule to use for selecting splitting
sizes. The final target distribution is the same regardless of splitting
schedule but the intermediate distributions can change. Current options
include
\itemize{
\item \code{split_district_only} At each step split off a single district.
\item \code{any_valid_sizes} At each step allow for regions to be split
into any two sizes (assuming the sizes can eventually be split into
districts.) Currently this is only supported for single member districting.
}
}
Parameters for \code{split_method} Any relevant parameters for the
\code{split_method}. These include the following
\itemize{
\item naive_top_k parameters:
\itemize{
\item \code{adapt_k_thresh} The threshold value used in the heuristic to
select a value \code{k_i} for each splitting iteration for graph space
sampling if estimation is desired (the \code{k_i} can also be manually passed in.)
Higher values are more accurate but may require more
computation. Set to 1 for the most conservative sampling.
Must be between 0 and 1.
\item \code{manual_k_params} The \code{k_i} values to be used for each splitting
iteration for graph space sampling. Beware when specifying manual values
it is crucial they are close to the true values as too small \code{k_i} values
will cause the algorithm to fail to sample from the target distribution
correctly and too large values will cause a drastic performance hit. The
input must either be a single integer to use for each step or a vector
of integers equal to the number of smc steps.
}
\item expo_bigger_abs_dev
\itemize{
\item \code{splitting_alpha} When selecting an edge to cut in the
tree a valid edge is selected with probability proportional to
\code{exp(-splitting_alpha * max_dev)}. \code{splitting_alpha} can
be any real number. Values closer to zero result in more stable weights
and larger values result in more unstable weights.
}
}}
}
\value{
A list of splitting parameters which is safe to pass into c++ code
}
\description{
Checks that the sampling space/splitting method combination given is valid
and that the splitting parameters are valid.
}
