% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walnuts.R
\name{walnuts_blk}
\alias{walnuts_blk}
\title{Selects Precincts and Blocks to Move to Equalize Population Between Districts}
\usage{
walnuts_blk(map_prec, map_blk, plan, tolerance = 5)
}
\arguments{
\item{map_prec}{a \link{redist_map} that is made up of precincts}

\item{plan}{an integer vector containing the plan to be equalized with one
value per precinct}

\item{tolerance}{an integer value that indicates at what deviation the
algorithm should no longer attempt to find a new set of district swaps (this
is only used very narrowly, does not apply to the whole algorithm)}

\item{map}{a \link{redist_map} that is made up of blocks}
}
\value{
a dataframe that contains the GEOID of every block, the GEOID of the
precinct that the block belongs to, and the district that it is assigned to
}
\description{
This function, given a map and a plan, finds a set of precincts and blocks
to move such that it equalizes the population between the districts defined
in the plan. It returns the new assignment of blocks and precincts to a plan.
}
\examples{
data(iowa)
iowa <- redist_map(iowa, existing_plan = cd_2010)
iowa_blk <- get_block_map("IA", 4)
#new_plan <- suppressWarnings(walnuts_blk(iowa, iowa_blk, iowa$cd_2010))

}
\concept{analyze}
