% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as.matrix.redist}
\alias{as.matrix.redist}
\title{Extract the redistricting matrix from a \code{redist} object}
\usage{
\method{as.matrix}{redist}(x, ...)
}
\arguments{
\item{x}{redist object}

\item{\dots}{additional arguments}
}
\value{
matrix of district assignments
}
\description{
Extract the redistricting matrix from a \code{redist} object
}
