% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{avg_by_prec}
\alias{avg_by_prec}
\title{Average a variable by precinct}
\usage{
avg_by_prec(plans, x, draws = NA)
}
\arguments{
\item{plans}{a `redist_plans` object}

\item{x}{an expression to average. Tidy-evaluated in `plans`.}

\item{draws}{which draws to average. `NULL` will average all draws, including
reference plans. The special value `NA` will average all sampled draws. An
integer, logical, or character vector indicating specific draws may also be
provided.}
}
\value{
a vector of length matching the number of precincts, containing the average.
}
\description{
Takes a column of a `redist_plans` object and averages it across a set of
`draws` for each precinct.
}
\concept{analyze}
