% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans_helpers.R
\name{avg_by_prec}
\alias{avg_by_prec}
\title{Average a variable by precinct (Deprecated)}
\usage{
avg_by_prec(plans, x, draws = NA)
}
\arguments{
\item{plans}{a \code{redist_plans} object}

\item{x}{an expression to average. Tidy-evaluated in \code{plans}.}

\item{draws}{which draws to average. \code{NULL} will average all draws, including
reference plans. The special value \code{NA} will average all sampled draws. An
integer, logical, or character vector indicating specific draws may also be
provided.}
}
\value{
a vector of length matching the number of precincts, containing the average.
}
\description{
Deprecated in favor of \code{\link[=proj_avg]{proj_avg()}}.
Takes a column of a \code{redist_plans} object and averages it across a set of
\code{draws} for each precinct.
}
