% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_calculations.R
\name{compute_log_optimal_weights}
\alias{compute_log_optimal_weights}
\title{Computes the unnormalized log target density for a ONE INDEXED! plans matrix
gSMC Redistricting Sampler (O'Sullivan, McCartan and Imai ???)}
\usage{
compute_log_optimal_weights(
  map,
  plans,
  counties = NULL,
  sizes_matrix = NULL,
  constraints = list(),
  splitting_schedule = "any_valid_sizes",
  num_threads = 0,
  compactness = 1L,
  pop_temper = 0L
)
}
\arguments{
\item{num_threads}{The number of threads used for computing
the target densities.}

\item{plan_matrix}{A matrix of 1-indexed plans}

\item{init_region_sizes_mat}{A matrix of region sizes. The matrix must have
the same number of columns as \code{plan_matrix}, \code{ndists} rows, and each column
must sum to \code{ndists}.}
}
\value{
A vector of the unnormalized target density of each plan
}
\description{
\code{generic_redist_gsmc} uses a Sequential Monte Carlo algorithm (O'Sullivan, McCartan and Imai ???)
to generate representative samples of congressional or legislative
redistricting plans according to contiguity, population, compactness, and
administrative boundary constraints.
}
\details{
Sampling can be performed either on the space of plans (graph partitions) or
on the space of spanning forests. This is an internal function designed to
support either sampling space and any arbitrary tree splitting method. The
functions \code{redist_gsmc} and \code{treedist_gsmc} are largely wrappers to calling
this function.

This function draws samples from a specific target measure controlled by
the \code{map} parameters.
}
