% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_calculations.R
\name{compute_log_target_density}
\alias{compute_log_target_density}
\title{Computes the unnormalized log target density for plans}
\usage{
compute_log_target_density(
  map,
  plans,
  counties = NULL,
  sizes_matrix = NULL,
  compactness = 1L,
  constraints = list(),
  num_threads = 0,
  pop_temper = 0L
)
}
\arguments{
\item{map}{A \code{\link[=redist_map]{redist_map()}} object.}

\item{plans}{Either a \code{redist_plans} object or a matrix or vector of plan labels}

\item{counties}{A vector containing county (or other administrative or
geographic unit) labels for each unit, which may be integers ranging from 1
to the number of counties, or a factor or character vector.  If provided,
the algorithm will only generate maps which split up to \code{ndists-1}
counties. Even if there are fewer counties than \code{ndists - 1}, the spanning
trees will change the results of the simulations. There is no strength
parameter associated with this constraint. To adjust the number of county
splits further, or to constrain a second type of administrative split,
consider using \code{add_constr_splits()}, \code{add_constr_multisplits()}, and
\code{add_constr_total_splits()}.}

\item{sizes_matrix}{A matrix of region sizes. The matrix must have
the same number of columns as \code{plans}, \code{ndists} rows, and each column
must sum to \code{ndists}.}

\item{compactness}{Controls the compactness of the generated districts, with
higher values preferring more compact districts. Must be nonnegative. See
the 'Details' section for more information, and computational
considerations.}

\item{constraints}{A \code{\link[=redist_constr]{redist_constr()}} object or a list containing
information on sampling constraints. See \link{constraints} for more information.}

\item{num_threads}{The number of threads used for computing
the target densities.}

\item{pop_temper}{The strength of the automatic population tempering. Try
values of 0.01-0.05 to start if the algorithm gets stuck on the final few
splits.}
}
\value{
A vector of the unnormalized target density of each plan
}
\description{
Computes the unnormalized log target density for the standard redist
target measure. The density for a plan  can be decomposed into a plan-term
(if relevant) and the procuct over regions of the regions compactness and
its score
}
