% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_splitting.R
\name{draw_tree_on_region}
\alias{draw_tree_on_region}
\title{Draw a spanning tree on a region of a plan}
\usage{
draw_tree_on_region(
  map,
  counties,
  partial_plan_labels,
  region_sizes,
  region_id_to_draw_tree_on,
  verbose
)
}
\arguments{
\item{map}{A \code{\link[=redist_map]{redist_map()}} object.}

\item{counties}{A vector containing county (or other administrative or
geographic unit) labels for each unit, which may be integers ranging from 1
to the number of counties, or a factor or character vector.  If provided,
the algorithm will only generate spanning trees where removing an edge splits
at most one county.}

\item{partial_plan_labels}{A vector or matrix of plan id labels. It should be
zero indexed. If its a matrix then only the first column is used.}

\item{region_sizes}{A vector or matrix of region sizes. If its a matrix then
only the first column is used.}

\item{region_id_to_draw_tree_on}{The id of the region to draw a tree on}

\item{verbose}{Whether or not to print the plan before a tree is drawn on it}
}
\value{
A list with the following
\itemize{
\item \code{uncut_tree} - he spanning tree drawn stored as a 0-indexed directed
adjacency list.
\item \code{root} - The 0-indexed root of the tree.
\item \code{num_attempts} -  The number of attempts it took to draw the tree.
\item \code{pop_below} - The population below each vertex in \code{uncut_tree} ie
the population induced by removing the edge terminating in that vertex
\item \code{uncut_tree_vertex_parents}: The parents of each zero-indexed vertex in
the tree. A value of -1 means that vertex is the root and -2 means that
vertex is not in the tree.
}
}
\description{
Draws a spanning tree uniformly at random on a region of a plan.
}
