% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_map.R
\name{get_adj}
\alias{get_adj}
\alias{set_adj}
\title{Get and set the adjacency graph from a \code{redist_map} object}
\usage{
get_adj(x)

set_adj(x, adj)
}
\arguments{
\item{x}{the \code{redist_map} object}

\item{adj}{a new adjacency list.}
}
\value{
a zero-indexed adjacency list (\code{get_adj})

the modified \code{redist_map} object (\code{set_adj})
}
\description{
Get and set the adjacency graph from a \code{redist_map} object
}
\concept{prepare}
