% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_plans.R
\name{get_plan_matrix}
\alias{get_plan_matrix}
\title{Extract the matrix of district assignments from a redistricting simulation}
\usage{
get_plan_matrix(plans)
}
\arguments{
\item{plans}{the \code{redist_plans} object}
}
\description{
Extract the matrix of district assignments from a redistricting simulation
}
