% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_plans.R
\name{get_plan_weights}
\alias{get_plan_weights}
\title{Extract the sampling weights from a redistricting simulation}
\usage{
get_plan_weights(plans)
}
\arguments{
\item{plans}{the \code{redist_plans} object}
}
\value{
the weights, with an additional attribute \code{resampled}
  indicating whether the plans have been resampled according to these
  weights.
}
\description{
Extract the sampling weights from a redistricting simulation
}
\concept{analyze}
