% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_map.R
\name{get_ref_plan_and_seats}
\alias{get_ref_plan_and_seats}
\title{Extract the reference plan from a \code{redist_map} object. If the map was
previously subsetted the districts in this plan will be labelled 1 to \code{ndists}}
\usage{
get_ref_plan_and_seats(x)
}
\arguments{
\item{x}{the \code{redist_map} object}
}
\value{
a vector of precinct assignments if a reference plan exists, if not
returns \code{NULL}
}
\description{
Extract the reference plan from a \code{redist_map} object. If the map was
previously subsetted the districts in this plan will be labelled 1 to \code{ndists}
}
