% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{pl}
\alias{pl}
\title{Access the Current \code{redist_plans()} Object}
\usage{
pl()
}
\value{
A \code{redist_plans} object, or \code{NULL} if not called from inside a
\code{dplyr} function.
}
\description{
Useful inside piped experessions and \code{dplyr} functions.
}
\examples{
pl()

}
