% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_map.R
\name{plot.redist_map}
\alias{plot.redist_map}
\title{Plot a \code{redist_map}}
\usage{
\method{plot}{redist_map}(x, fill = NULL, ...)
}
\arguments{
\item{x}{the \code{redist_map} object}

\item{fill}{\code{\link[dplyr:dplyr_data_masking]{<data-masking>}} If absent,
\code{\link{redist.map}} will be called; if present, will be used to color
the map units using \code{\link{redist.choropleth}}. If using data masking,
may need to explicitly name argument \code{fill=...} to avoid S3 generic
issues.}

\item{...}{passed on to the underlying functions.}
}
\description{
Plot a \code{redist_map}
}
\examples{
data(fl25)
d = redist_map(fl25, ndists=3, pop_tol=0.05)
plot(d)
plot(d, BlackPop/pop)

data(fl25_enum)
fl25$dist <- fl25_enum$plans[, 5118]
d <- redist_map(fl25, existing_plan = dist)
plot(d)



}
\concept{plot}
\concept{prepare}
