% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_plans.R
\name{plot.redist_plans}
\alias{plot.redist_plans}
\title{Summary plots for \code{redist_plans}}
\usage{
\method{plot}{redist_plans}(x, ..., type = "hist")
}
\arguments{
\item{x}{the \code{redist_plans} object.}

\item{...}{passed on to the underlying function}

\item{type}{the name of the plotting function to use. Will have \code{plot_},
prepended to it; e.g., use \code{type="plan"} to call \code{plot_plan()}.}
}
\description{
Summary plots for \code{redist_plans}
}
\concept{plot}
