% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_plans.R
\name{plot.redist_plans}
\alias{plot.redist_plans}
\title{Summary plots for \code{\\link{redist_plans}}}
\usage{
\method{plot}{redist_plans}(x, ..., type = "distr_qtys")
}
\arguments{
\item{x}{the \code{redist_plans} object.}

\item{...}{passed on to the underlying function}

\item{type}{the name of the plotting function to use. Will have
\code{redist.plot.}, prepended to it; e.g., use \code{type="plans"} to call
\code{\link{redist.plot.plans}}.}
}
\description{
If no arguments are passed, defaults to plotting the sampling weights for
the \code{\link{redist_plans}} object. If no weights exist, plots district
populations.
}
\concept{plot}
