% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{prec_assignment}
\alias{prec_assignment}
\title{Extract the district assignments for a precinct across all simulated plans}
\usage{
prec_assignment(prec, .data = cur_plans())
}
\arguments{
\item{prec}{the precinct number or ID}

\item{.data}{a \code{\link{redist_plans}} object}
}
\value{
integer vector, a row from a plans matrix
}
\description{
Extract the district assignments for a precinct across all simulated plans
}
\concept{analyze}
