% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans_helpers.R
\name{pullback}
\alias{pullback}
\title{Pull back plans to unmerged units}
\usage{
pullback(plans, map = NULL)
}
\arguments{
\item{plans}{a \code{redist_plans} object}

\item{map}{optionally, a \code{redist_map} object, which will be used to set the new population vector}
}
\value{
a new, re-indexed, \code{redist_plans} object
}
\description{
Merging map units through \code{\link{merge_by}} or \code{\link{summarize}}
changes the indexing of each unit.  Use this function to take a set of
redistricting plans from a \code{redist} algorithm and re-index them to
be compatible with the original set of units.
}
\concept{analyze}
