% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{redist.adjacency}
\alias{redist.adjacency}
\title{Adjacency List functionality for redist}
\usage{
redist.adjacency(shp, plan)
}
\arguments{
\item{shp}{A SpatialPolygonsDataFrame or sf object. Required.}

\item{plan}{A numeric vector (if only one map) or matrix with one row}
}
\value{
Adjacency list
}
\description{
Creates an adjacency list that is zero indexed with no skips
}
