% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{redist.choropleth}
\alias{redist.choropleth}
\title{Creates a Choropleth}
\usage{
redist.choropleth(
  shp,
  fill = NULL,
  fill_label = "",
  title = "",
  limit_colors = NULL,
  grad = 1,
  lwd = 0
)
}
\arguments{
\item{shp}{A SpatialPolygonsDataFrame or sf object. Required.}

\item{fill}{A numeric/integer vector with values to color the plot with. Optional.}

\item{fill_label}{A string title of plot. Defaults to empty string. Optional.}

\item{title}{A string title of plot. Defaults to empty string. Optional.}

\item{limit_colors}{A length two string vector with two colors, either hex or ggplot color names.}

\item{grad}{Number of colors to make a gradient with. Accepts values of 1 or 2.}

\item{lwd}{Line width. Defaults to 0}
}
\value{
ggplot map
}
\description{
Creates a Choropleth
}
\details{
Creates a basic choropleth for a provided shp with value. Recommended to
normalize data to avoid absolute values, in most use cases.
}
\examples{
\dontrun{
data("fl25")
redist.choropleth(shp = fl25, fill = BlackPop/TotPop)
DVS <- fl25$obama/(fl25$mccain+fl25$obama)
redist.choropleth(shp = fl25, fill = DVS, limit_colors = c('red', 'blue'))
}

}
\concept{plot}
