% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist.R
\name{redist.combine.anneal}
\alias{redist.combine.anneal}
\title{redist.combine.anneal}
\usage{
redist.combine.anneal(file_name)
}
\arguments{
\item{file_name}{The file name to search for in current working directory.}
}
\description{
Combine files generated by redist.mcmc.anneal()
}
