% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitiveness.R, R/tidy.R
\name{redist.competitiveness}
\alias{redist.competitiveness}
\alias{competitiveness}
\title{Compute Competitiveness}
\usage{
redist.competitiveness(district_membership, rvote, dvote)

competitiveness(map, rvote, dvote, .data = get0(".", parent.frame()))
}
\arguments{
\item{district_membership}{A numeric vector (if only one map) or matrix with one row 
for each precinct and one column for each map. Required.}

\item{rvote}{A numeric vector with the Republican vote for each precinct.}

\item{dvote}{A numeric vector with the Democratic vote for each precinct.}

\item{map}{a \code{\link{redist_map}} object}

\item{.data}{a \code{\link{redist_plans}} object}
}
\value{
Numeric vector with competitiveness scores
}
\description{
Currently only implements the competitiveness function in equation (5) 
of Cho & Liu 2016.
}
\examples{
\dontrun{
data("algdat.p10")
comp <- redist.competitiveness(algdat.p10$cdmat, 
algdat.p10$precinct.data$repvote, 
algdat.p10$precinct.data$demvote)
}
}
\concept{analyze}
