% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counties.R
\name{redist.county.relabel}
\alias{redist.county.relabel}
\title{Relabel Discontinuous Counties}
\usage{
redist.county.relabel(adjacency, counties)
}
\arguments{
\item{adjacency}{adjacency list}

\item{counties}{character vector of county names}
}
\value{
character vector of county names
}
\description{
Relabel Discontinuous Counties
}
\examples{
\dontrun{

set.seed(2)
counties <- sample(c(rep('a', 20), rep('b', 5)))
redist.relabel.counties(counties)

}
}
