% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counties.R
\name{redist.county.relabel}
\alias{redist.county.relabel}
\title{Relabel Discontinuous Counties}
\usage{
redist.county.relabel(adj, counties, simplify = TRUE)
}
\arguments{
\item{adj}{adjacency list}

\item{counties}{character vector of county names}

\item{simplify}{boolean - TRUE returns a numeric vector of ids, while FALSE
appends a number when there are multiple connected components.}
}
\value{
character vector of county names
}
\description{
Relabel Discontinuous Counties
}
\examples{
set.seed(2)
data(fl25)
data(fl25_adj)
counties <- sample(c(rep('a', 20), rep('b', 5)))
redist.county.relabel(fl25_adj, counties)

}
\concept{prepare}
