% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumpart.R
\name{redist.enumpart}
\alias{redist.enumpart}
\title{Enumerate All Partitions (Fifield et al. 2020)}
\usage{
redist.enumpart(
  adj,
  ordered_path,
  out_path,
  ndists = 2,
  all = TRUE,
  n = NULL,
  weight_path = NULL,
  lower = NULL,
  upper = NULL,
  init = FALSE,
  read = TRUE,
  total_pop = NULL,
  unordered_path
)
}
\arguments{
\item{adj}{zero indexed adjacency list.}

\item{ordered_path}{valid path to output the ordered adjacency map to}

\item{out_path}{Valid path to output the enumerated districts}

\item{ndists}{number of districts to enumerate}

\item{all}{boolean. TRUE outputs all districts. FALSE samples n districts.}

\item{n}{integer. Number of districts to output if all is FALSE. Returns
districts selected from uniform random distribution.}

\item{weight_path}{A path (not including ".dat") to a space-delimited file containing a vector of
vertex weights, to be used along with \code{lower} and \code{upper}.}

\item{lower}{A lower bound on each partition's total weight, implemented by rejection sampling.}

\item{upper}{An upper bound on each partition's total weight.}

\item{init}{Runs redist.init.enumpart. Defaults to FALSE. Should be run on first use.}

\item{read}{boolean. Defaults to TRUE. If TRUE, reads and returns the results.}

\item{total_pop}{the vector of precinct populations}

\item{unordered_path}{Deprecated.}
}
\value{
List with entries district_membership and parity.
}
\description{
Single function for standard enumeration analysis, using ZDD methodology
(Fifield, Imai, Kawahara, and Kenny 2020).
}
\references{
Fifield, B., Imai, K., Kawahara, J., & Kenny, C. T. (2020). The essential
role of empirical validation in legislative redistricting simulation.
\emph{Statistics and Public Policy}, 7(1), 52-68.
}
\concept{enumerate}
