% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumpart.R
\name{redist.enumpart}
\alias{redist.enumpart}
\title{Enumerate All Partitions}
\usage{
redist.enumpart(
  adj,
  ordered_path,
  out_path,
  ndists = 2,
  all = TRUE,
  n = NULL,
  weight_path = NULL,
  lower = NULL,
  upper = NULL,
  init = FALSE,
  read = TRUE,
  total_pop = NULL,
  unordered_path
)
}
\arguments{
\item{adj}{Zero-indexed adjacency list.}

\item{ordered_path}{Valid path to output the ordered adjacency map to.}

\item{out_path}{Valid path to output the enumerated districts.}

\item{ndists}{Number of districts to enumerate.}

\item{all}{If \code{TRUE}, outputs all districts. If \code{FALSE}, draws \code{n}
districts uniformly at random.}

\item{n}{Number of districts to output if \code{all} is \code{FALSE}.}

\item{weight_path}{A path (not including \code{".dat"}) to a vertex weights file.}

\item{lower}{A lower bound on each partition's total weight.}

\item{upper}{An upper bound on each partition's total weight.}

\item{init}{If \code{TRUE}, runs initialization first.}

\item{read}{If \code{TRUE}, reads and returns the results.}

\item{total_pop}{The vector of precinct populations.}

\item{unordered_path}{Deprecated, ignored.}
}
\value{
A list with \code{plans} (district membership matrix) and \code{parity}.
}
\description{
\strong{Deprecated}: Use \code{\link[=redist_enumpart]{redist_enumpart()}} instead.

Single function for standard enumeration analysis, using ZDD methodology
(Fifield, Imai, Kawahara, and Kenny 2020).
}
\references{
Fifield, B., Imai, K., Kawahara, J., & Kenny, C. T. (2020).
The essential role of empirical validation in legislative redistricting
simulation. \emph{Statistics and Public Policy}, 7(1), 52-68.
}
\concept{enumerate}
\keyword{internal}
