% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findtarget.R
\name{redist.find.target}
\alias{redist.find.target}
\title{Find Majority Minority Remainder}
\usage{
redist.find.target(tgt_min, group_pop, total_pop, ndists, nmmd)
}
\arguments{
\item{tgt_min}{target group population for majority minority district}

\item{group_pop}{A vector of populations for some subgroup of interest.}

\item{total_pop}{A vector containing the populations of each geographic unit.}

\item{ndists}{The number of congressional districts.}

\item{nmmd}{The number of majority minority districts.}
}
\value{
numeric value to target
}
\description{
Given a percent goal for majority minority districts, this computes the average
value of minority in non-majority minority districts. This value is "tgt_other"
in \code{redist.flip} and \code{redist_smc}.
}
\concept{prepare}
