% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freeze.R
\name{redist.freeze}
\alias{redist.freeze}
\alias{freeze}
\title{Freeze Parts of a Map}
\usage{
redist.freeze(adj, plan, freeze_row)

freeze(freeze_row, adj, plan, .data = get0(".", parent.frame()))
}
\arguments{
\item{adj}{Required, zero indexed adjacency list.}

\item{plan}{Recommended, whether to create separate groups by district.}

\item{freeze_row}{Required, logical vector where TRUE freezes and FALSE lets a precinct 
stay free or a vector of indices to freeze}

\item{.data}{a \code{\link{redist_map}} object}
}
\value{
integer vector to group by
}
\description{
Freeze Parts of a Map
}
\examples{
  data(fl25)
  data(algdat.p10)
adj <- redist.adjacency(fl25)
freeze_id <- redist.freeze(adj = adj, plan = algdat.p10$cdmat[, 1], 
freeze_row =  (2 == algdat.p10$cdmat[, 1]))
}
\concept{prepare}
