% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_percent.R
\name{redist.group.percent}
\alias{redist.group.percent}
\title{Calculate Group Percent by District}
\usage{
redist.group.percent(
  plans,
  district_membership,
  group_pop,
  grouppop,
  total_pop,
  fullpop,
  ncores = 1
)
}
\arguments{
\item{plans}{A matrix with one row 
for each precinct and one column for each map. Required.}

\item{district_membership}{Deprecated, use plans. A matrix with one row 
for each precinct and one column for each map. Required.}

\item{group_pop}{A numeric vector with the population of the group for every precinct.}

\item{grouppop}{Deprecated, use group_pop. A numeric vector with the population of the group for every precinct.}

\item{total_pop}{A numeric vector with the population for every precinct.}

\item{fullpop}{Deprecated, use total_pop. A numeric vector with the population for every precinct.}

\item{ncores}{Number of cores to use for parallel computing. Default is 1.}
}
\value{
matrix with percent for each district
}
\description{
\code{redist.group.percent} computes the percentage that a group makes up in
each district across a matrix of maps.
}
\examples{
\dontrun{
data("fl25")
data(algdat)
cd <- algdat.p10$cdmat[,1:5]

redist.group.percent(plans = cd, 
                    group_pop = fl25$BlackPop, 
                    total_pop = fl25$TotPop)
}
}
