% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_percent.R
\name{redist.group.percent}
\alias{redist.group.percent}
\title{Calculate Group Percent by District}
\usage{
redist.group.percent(district_membership, grouppop, fullpop, ncores = 1)
}
\arguments{
\item{district_membership}{A matrix with one row 
for each precinct and one column for each map. Required.}

\item{grouppop}{A numeric vector with the population of the group for every precinct.}

\item{fullpop}{A numeric vector with the population for every precinct.}

\item{ncores}{Number of cores to use for parallel computing. Default is 1.}
}
\value{
matrix with percent for each district
}
\description{
\code{redist.group.percent} computes the percentage that a group makes up in
each district across a matrix of maps.
}
\examples{
\dontrun{
data("fl25")
data(algdat)
cd <- algdat.p10$cdmat[,1:5]

redist.group.percent(district_membership = cd, 
                    grouppop = fl25$BlackPop, 
                    fullpop = fl25$TotPop)
}
}
