% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_percent.R, R/tidy.R
\name{redist.group.percent}
\alias{redist.group.percent}
\alias{group_frac}
\title{Calculate Group Percent by District}
\usage{
redist.group.percent(
  plans,
  group_pop,
  total_pop,
  ncores = 1,
  district_membership,
  grouppop,
  fullpop
)

group_frac(
  map,
  group_pop,
  total_pop = map[[attr(map, "pop_col")]],
  .data = get0(".", parent.frame())
)
}
\arguments{
\item{plans}{A matrix with one row
for each precinct and one column for each map. Required.}

\item{group_pop}{A numeric vector with the population of the group for every precinct.}

\item{total_pop}{A numeric vector with the population for every precinct.}

\item{ncores}{Number of cores to use for parallel computing. Default is 1.}

\item{district_membership}{Deprecated, use plans. A matrix with one row
for each precinct and one column for each map. Required.}

\item{grouppop}{Deprecated, use group_pop. A numeric vector with the population of the group for every precinct.}

\item{fullpop}{Deprecated, use total_pop. A numeric vector with the population for every precinct.}

\item{map}{a \code{\link{redist_map}} object}

\item{.data}{a \code{\link{redist_plans}} object}
}
\value{
matrix with percent for each district
}
\description{
\code{redist.group.percent} computes the percentage that a group makes up in
each district across a matrix of maps.
}
\examples{
\dontrun{
data("fl25")
data(algdat)
cd <- algdat.p10$cdmat[,1:5]

redist.group.percent(plans = cd,
                    group_pop = fl25$BlackPop,
                    total_pop = fl25$TotPop)
}
}
\concept{analyze}
