% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_percent.R, R/tidy.R
\name{redist.group.percent}
\alias{redist.group.percent}
\alias{group_frac}
\title{Calculate Group Percent by District}
\usage{
redist.group.percent(district_membership, group_pop, full_pop)

group_frac(
  map,
  group_pop,
  full_pop = map[[attr(map, "pop_col")]],
  .data = get0(".", parent.frame())
)
}
\arguments{
\item{district_membership}{A matrix with one row
for each precinct and one column for each map. Required.}

\item{group_pop}{A numeric vector with the population of the group for every precinct.}

\item{full_pop}{A numeric vector with the population for every precinct.}

\item{map}{a \code{\link{redist_map}} object}

\item{.data}{a \code{\link{redist_plans}} object}
}
\value{
A matrix with a row for each district and a column for each plan,
 with the entries representing the percentage of the group in that district,
 relative to the total population of the district.
}
\description{
\code{redist.group.percent} computes the percentage that a group makes up in
each district across a matrix of maps.
}
\examples{
\dontrun{
data("fl25")
data(algdat)
cd <- algdat.p10$cdmat[,1:5]

redist.group.percent(district_membership = cd,
                    grouppop = fl25$BlackPop,
                    fullpop = fl25$TotPop)
}
}
\concept{analyze}
