% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cores.R, R/tidy.R
\name{redist.identify.cores}
\alias{redist.identify.cores}
\alias{make_cores}
\title{Identify Cores of a District (Heuristic)}
\usage{
redist.identify.cores(
  adj,
  plan,
  within = 1,
  focus = NULL,
  simplify = TRUE,
  adjacency,
  district_membership
)

make_cores(.data = get0(".", parent.frame()), within = 1, focus = NULL)
}
\arguments{
\item{adj}{zero indexed adjacency list.}

\item{plan}{An integer vector or matrix column of district assignments.}

\item{within}{the core is defined to be at least this number of steps within
district boundaries}

\item{focus}{Optional. Integer. A single district to focus on.}

\item{simplify}{Optional. Logical. Whether to return extra information or just grouping ID.}

\item{adjacency}{Deprecated, use adj. A zero indexed adjacency list.}

\item{district_membership}{Deprecated, use plan. An integer vector or matrix column of district assignments.}

\item{.data}{a \code{\link{redist_map}} object}
}
\value{
integer vector if simplify is false. Otherwise returns a tibble with the grouping
variable and additional information.
}
\description{
Creates a grouping ID to unite geographies and perform analysis on a smaller
set of precincts. It identifies all precincts more than \code{within} edges
of a district district border Each precinct more than \code{within} of
another district gets it own group.
}
\concept{prepare}
