% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cores.R
\name{redist.identify.cores}
\alias{redist.identify.cores}
\title{Identify Cores of a District (Heuristic)}
\usage{
redist.identify.cores(
  adj,
  adjacency,
  plan,
  district_membership,
  k = 1,
  focus = NULL,
  simplify = TRUE
)
}
\arguments{
\item{adj}{zero indexed adjacency list.}

\item{adjacency}{Deprecated, use adj. A zero indexed adjacency list.}

\item{plan}{An integer vector or matrix column of district assignments.}

\item{district_membership}{Deprecated, use plan. An integer vector or matrix column of district assignments.}

\item{k}{Number of steps to check for. Defaults to 1.}

\item{focus}{Optional. Integer. A single district to focus on.}

\item{simplify}{Optional. Logical. Whether to return extra information or just grouping ID.}
}
\value{
integer vector if simplify is false. Otherwise returns a tibble with the grouping 
variable and additional information.
}
\description{
Creates a grouping ID to unite geographies and perform analysis on a smaller set of
precincts. Given a k, it identifies all precincts within k edges of another district
on a graph. Each precinct within k of another district gets it own group. Connected
components more than k away are givent the same grouping variable.
}
