% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumpart.R
\name{redist.init.enumpart}
\alias{redist.init.enumpart}
\title{Initialize enumpart}
\usage{
redist.init.enumpart()
}
\value{
0 on success.
}
\description{
\strong{Deprecated}: Use \code{\link[=redist_enumpart]{redist_enumpart()}} with \code{init = TRUE} instead.
}
\references{
Fifield, B., Imai, K., Kawahara, J., & Kenny, C. T. (2020).
The essential role of empirical validation in legislative redistricting
simulation. \emph{Statistics and Public Policy}, 7(1), 52-68.
}
\concept{enumerate}
\keyword{internal}
