% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumpart.R
\name{redist.init.enumpart}
\alias{redist.init.enumpart}
\title{Initialize enumpart}
\usage{
redist.init.enumpart()
}
\value{
0 on success
}
\description{
This ensures that the enumerate partitions programs is prepared to run.
This must be run once per install of the redist package.
}
\examples{
\dontrun{
redist.init.enumpart()
}
}
\references{
Benjamin Fifield, Kosuke Imai, Jun Kawahara, and Christopher T Kenny.
"The Essential Role of Empirical Validation in Legislative Redistricting Simulation."
Forthcoming, Statistics and Public Policy.
}
\concept{enumerate}
