% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parity.R
\name{redist.parity}
\alias{redist.parity}
\alias{plan_parity}
\title{Calculates Population Parity}
\usage{
redist.parity(plans, total_pop, ncores = 1, district_membership, population)

plan_parity(map, .data = get0(".", parent.frame()), ...)
}
\arguments{
\item{plans}{A matrix with one row for each precinct and one column for each
map. Required.}

\item{total_pop}{A numeric vector with the population for every precinct.}

\item{ncores}{Number of cores to use for parallel computing. Default is 1.}

\item{district_membership}{Deprecated, use plans. A matrix with one row
for each precinct and one column for each map. Required.}

\item{population}{Deprecated, use total_pop. A numeric vector with the population for every precinct.}

\item{map}{a \code{\link{redist_map}} object}

\item{.data}{a \code{\link{redist_plans}} object}

\item{...}{passed on to \code{redist.parity}}
}
\value{
numeric vector with the population parity for each column
}
\description{
\code{redist.parity} computes the population parity of a matrix of maps.
}
\concept{analyze}
