% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_plans.R
\name{redist.plot.hist}
\alias{redist.plot.hist}
\alias{hist.redist_plans}
\title{Plot a histogram of a summary statistic}
\usage{
redist.plot.hist(x, qty, bins = NULL, ...)

\method{hist}{redist_plans}(x, qty, ...)
}
\arguments{
\item{x}{the \code{redist_plans} object.}

\item{qty}{\code{\link[dplyr:dplyr_data_masking]{<data-masking>}} the statistic.}

\item{bins}{the number of bins to use in the histogram. Defaults to Freedman-Diaconis rule.}

\item{...}{passed on to \code{\link[ggplot2]{geom_histogram}}}
}
\description{
Plots a histogram of a statistic of a \code{\link{redist_plans}} object,
with a reference line for each reference plan, if applicable.
}
\concept{plot}
