% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_penalty.R
\name{redist.plot.penalty}
\alias{redist.plot.penalty}
\title{(Deprecated) Visualize Group Power Penalty}
\usage{
redist.plot.penalty(
  tgt_min = 0.55,
  tgt_other = 0.25,
  strength_vra = 2500,
  pow_vra = 1.5,
  limits = TRUE
)
}
\arguments{
\item{tgt_min}{double, defaults to 0.55. The minority target percent.}

\item{tgt_other}{double, defaults to 0.25. The other group target percent.}

\item{strength_vra}{double, strength of the VRA constraint.}

\item{pow_vra}{double, exponent of the VRA constraint.}

\item{limits}{Whether to limit y axis to 0,500. Default is TRUE for
comparability across values.}
}
\value{
ggplot
}
\description{
Plots the shape of the [add_constr_grp_pow()] penalty.
}
\details{
This function allows you to plot the un-exponentiated penalty implemented as
[add_constr_grp_pow()]. The function takes two key inputs,
`tgt_min` and `tgt_other` which center the minimum penalty spots. A higher y-value
indicates a higher penalty and incentivizes moving towards a spot with a lower y-value.
The x-axis indicates the group population proportion in a given district.
}
\concept{prepare}
