% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumpart.R
\name{redist.prep.enumpart}
\alias{redist.prep.enumpart}
\title{Prepares a run of the enumpart algorithm by ordering edges}
\usage{
redist.prep.enumpart(
  adj,
  ordered_path,
  weight_path = NULL,
  total_pop = NULL,
  unordered_path
)
}
\arguments{
\item{adj}{zero indexed adjacency list}

\item{ordered_path}{valid path to output the ordered adjacency map to}

\item{weight_path}{A path (not including ".dat") to store a space-delimited
file containing a vector of vertex weights. Only supply with total_pop.}

\item{total_pop}{the vector of precinct populations. Only supply with weight_path.}

\item{unordered_path}{Deprecated.}
}
\value{
0 on success
}
\description{
Prepares a run of the enumpart algorithm by ordering edges
}
\examples{
\dontrun{
temp <- tempdir()
data(fl25)
adj <- redist.adjacency(fl25)
redist.prep.enumpart(adj = adj, ordered_path = paste0(temp, "/ordered"))
}
}
\references{
Benjamin Fifield, Kosuke Imai, Jun Kawahara, and Christopher T Kenny.
"The Essential Role of Empirical Validation in Legislative Redistricting Simulation."
Forthcoming, Statistics and Public Policy.
}
\concept{enumerate}
