% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumpart.R
\name{redist.read.enumpart}
\alias{redist.read.enumpart}
\title{Read Results from enumpart}
\usage{
redist.read.enumpart(out_path, skip = 0, n_max = -1L)
}
\arguments{
\item{out_path}{\code{out_path} specified in \code{redist.run.enumpart()}.}

\item{skip}{Number of lines to skip.}

\item{n_max}{Maximum number of lines to read.}
}
\value{
A district membership matrix (1-indexed).
}
\description{
\strong{Deprecated}: Use \code{\link[=redist_enumpart]{redist_enumpart()}} instead, which reads results
automatically.
}
\references{
Fifield, B., Imai, K., Kawahara, J., & Kenny, C. T. (2020).
The essential role of empirical validation in legislative redistricting
simulation. \emph{Statistics and Public Policy}, 7(1), 52-68.
}
\concept{enumerate}
\keyword{internal}
