% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{redist.reduce.adjacency}
\alias{redist.reduce.adjacency}
\title{Reduce Adjacency List}
\usage{
redist.reduce.adjacency(adj, keep_rows)
}
\arguments{
\item{adj}{A zero-indexed adjacency list. Required.}

\item{keep_rows}{row numbers of precincts to keep}
}
\value{
zero indexed adjacency list with max value length(keep_rows) - 1
}
\description{
Tool to help reduce adjacency lists for analyzing subsets of maps.
}
\examples{
data(fl25_adj)
redist.reduce.adjacency(fl25_adj, c(2, 3, 4, 6, 21))

}
\concept{prepare}
