% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumpart.R
\name{redist.run.enumpart}
\alias{redist.run.enumpart}
\title{Run the enumpart algorithm}
\usage{
redist.run.enumpart(
  ordered_path,
  out_path,
  ndists = 2,
  all = TRUE,
  n = NULL,
  weight_path = NULL,
  lower = NULL,
  upper = NULL,
  options = NULL
)
}
\arguments{
\item{ordered_path}{Path used in \code{redist.prep.enumpart()} (not including
\code{".dat"}).}

\item{out_path}{Valid path to output the enumerated districts.}

\item{ndists}{Number of districts to enumerate.}

\item{all}{If \code{TRUE}, outputs all districts. If \code{FALSE}, draws \code{n}
districts uniformly at random.}

\item{n}{Number of districts to output if \code{all} is \code{FALSE}.}

\item{weight_path}{A path (not including \code{".dat"}) to a vertex weights file.}

\item{lower}{A lower bound on each partition's total weight.}

\item{upper}{An upper bound on each partition's total weight.}

\item{options}{Additional enumpart arguments. Not recommended.}
}
\value{
0 on success.
}
\description{
\strong{Deprecated}: Use \code{\link[=redist_enumpart]{redist_enumpart()}} instead.
}
\references{
Fifield, B., Imai, K., Kawahara, J., & Kenny, C. T. (2020).
The essential role of empirical validation in legislative redistricting
simulation. \emph{Statistics and Public Policy}, 7(1), 52-68.
}
\concept{enumerate}
\keyword{internal}
