% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate.R
\name{redist.samplepart}
\alias{redist.samplepart}
\title{Sample partitions using spanning trees}
\usage{
redist.samplepart(adjobj, ndists, contiguitymap, nsamp)
}
\arguments{
\item{adjobj}{An adjacency list, matrix, or object of class
\code{SpatialPolygonsDataFrame}.}

\item{ndists}{The desired number of congressional districts}

\item{contiguitymap}{Use queens or rooks distance criteria for generating an
adjacency list from a "SpatialPolygonsDataFrame" data type.
Default is "rooks".}

\item{nsamp}{Number of samples to draw before reweighting. Default is 1000.}
}
\value{
\code{redist.samplepart} returns a list where the first entry is the
randomly sampled redistricting plan, and the second entry is the number of
possible redistricting plans from the implied spanning tree.
}
\description{
\code{redist.samplepart} uses a spanning tree method to randomly sample
redistricting plans.
}
