% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcstats.R, R/tidy.R
\name{redist.segcalc}
\alias{redist.segcalc}
\alias{segregation_index}
\title{Segregation index calculation for MCMC redistricting.}
\usage{
redist.segcalc(plans, group_pop, total_pop, algout, grouppop, fullpop)

segregation_index(
  map,
  group_pop,
  total_pop = map[[attr(map, "pop_col")]],
  .data = get0(".", parent.frame())
)
}
\arguments{
\item{plans}{A matrix of congressional district assignments or a
redist object.}

\item{group_pop}{A vector of populations for some subgroup of interest.}

\item{total_pop}{A vector containign the populations of each geographic unit.}

\item{algout}{Deprecated. Use plans. A matrix of congressional district assignments or a
redist object.}

\item{grouppop}{Deprecated. Use group_pop. A vector of populations for some subgroup of interest.}

\item{fullpop}{Deprecated. Use total_pop. A vector containign the populations of each geographic unit.}

\item{map}{a \code{\link{redist_map}} object}

\item{.data}{a \code{\link{redist_plans}} object}
}
\value{
\code{redist.segcalc} returns a vector where each entry is the
dissimilarity index of segregation (Massey & Denton 1987) for each
redistricting plan in \code{algout}.
}
\description{
\code{redist.segcalc} calculates the dissimilarity index of segregation (see
Massey \& Denton 1987 for more details) for a specified subgroup under any
redistricting plan.
}
\examples{
\dontrun{
data(algdat.pfull)

## Code to run the simulations in Figure 4 of Fifield, Higgins,
## Imai and Tarr (2015)

## Get an initial partition
set.seed(1)
initcds <- algdat.pfull$cdmat[,sample(1:ncol(algdat.pfull$cdmat), 1)]

## Run simulations
alg_253 <- redist.mcmc(adjobj = algdat.pfull$adjlist,
popvec = algdat.pfull$precinct.data$pop,
initcds = initcds, nsims = 10000)

## Get Republican Dissimilarity Index from simulations
rep_dmi_253 <- redist.segcalc(alg_253,
algdat.pfull$precinct.data$repvote,
algdat.pfull$precinct.data$pop)
}
}
\references{
Fifield, Benjamin, Michael Higgins, Kosuke Imai and Alexander
Tarr. (2016) "A New Automated Redistricting Simulator Using Markov Chain
Monte Carlo." Working Paper. Available at
\url{http://imai.princeton.edu/research/files/redist.pdf}.

Massey, Douglas and Nancy Denton. (1987) "The Dimensions of Social
Segregation". Social Forces.
}
\concept{analyze}
