% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splits.R
\name{redist.splits}
\alias{redist.splits}
\title{Count County Splits}
\usage{
redist.splits(district_membership, counties)
}
\arguments{
\item{district_membership}{A numeric vector (if only one map) or matrix with one row 
for each precinct and one column for each map. Required.}

\item{counties}{A vector of county names or county ids.}
}
\value{
integer vector with one number for each map
}
\description{
Count County Splits
}
