% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R, R/splits.R
\name{county_splits}
\alias{county_splits}
\alias{redist.splits}
\title{Count County Splits}
\usage{
county_splits(map, counties, .data = cur_plans())

redist.splits(plans, counties)
}
\arguments{
\item{map}{a \code{\link{redist_map}} object}

\item{counties}{A vector of county names or county ids.}

\item{.data}{a \code{\link{redist_plans}} object}

\item{plans}{A numeric vector (if only one map) or matrix with one row
for each precinct and one column for each map. Required.}
}
\value{
integer vector with one number for each map
}
\description{
Count County Splits
}
\concept{analyze}
