% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_flip_tidy.R
\name{redist_flip}
\alias{redist_flip}
\title{Tidy Flip Markov Chain Monte Carlo}
\usage{
redist_flip(
  map,
  nsims,
  init_plan,
  counties = NULL,
  group_pop,
  constraints = list(),
  eprob = 0.05,
  lambda = 0,
  temper = FALSE,
  betaseq = "powerlaw",
  betaseqlength = 10,
  betaweights = NULL,
  adapt_lambda = FALSE,
  adapt_eprob = FALSE,
  exact_mh = FALSE,
  adjswaps = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{map}{A \code{\link{redist_map}} object.}

\item{nsims}{The number of samples to draw.}

\item{init_plan}{A vector containing the congressional district labels
of each geographic unit. The default is \code{NULL}. If not provided,
a random initial plan will be generated using \code{redist.smc}.}

\item{counties}{A column in map containing county membership}

\item{group_pop}{A column in map containing group populations}

\item{constraints}{a list of constraints to implement. Can be created with 
\code{flip_constraints_helper}}

\item{eprob}{The probability of keeping an edge connected. The
default is \code{0.05}.}

\item{lambda}{lambda The parameter determining the number of swaps to attempt
each iteration of the algorihtm. The number of swaps each iteration is
equal to Pois(\code{lambda}) + 1. The default is \code{0}.}

\item{temper}{Whether to use simulated tempering algorithm. Default is FALSE.}

\item{betaseq}{Sequence of beta values for tempering. The default is
\code{powerlaw} (see Fifield et. al (2020) for details).}

\item{betaseqlength}{@param betaseqlength Length of beta sequence desired for
tempering. The default is \code{10}.}

\item{betaweights}{betaweights Sequence of weights for different values of
beta. Allows the user to upweight certain values of beta over
others. The default is \code{NULL} (equal weighting).}

\item{adapt_lambda}{adapt_lambda Whether to adaptively tune the lambda parameter so that the Metropolis-Hastings
acceptance probability falls between 20\% and 40\%. Default is FALSE.}

\item{adapt_eprob}{eprob Whether to adaptively tune the edgecut probability parameter so that the
Metropolis-Hastings acceptance probability falls between 20\% and 40\%. Default is
FALSE.}

\item{exact_mh}{Whether to use the approximate (FALSE) or exact (TRUE)
Metropolis-Hastings ratio calculation for accept-reject rule. Default is FALSE.}

\item{adjswaps}{Flag to restrict swaps of beta so that only
values adjacent to current constraint are proposed. The default is
\code{TRUE}.}

\item{verbose}{Whether to print initialization statement. Default is \code{TRUE}.}
}
\value{
redist plans object
}
\description{
Tidy Flip Markov Chain Monte Carlo
}
\examples{
\dontrun{
data(fl25)
fl25_map <- redist_map(fl, ndists = 4, existing_plan = cd)
sims <- redist_flip(map = fl25_map, nsims = 10)
}
}
