% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_gsmc.R
\name{redist_gsmc}
\alias{redist_gsmc}
\title{gSMC Redistricting Sampler}
\usage{
redist_gsmc(
  state_map,
  M,
  counties = NULL,
  k_param_val = 6,
  resample = TRUE,
  runs = 1L,
  ncores = 0L,
  verbose = FALSE,
  silent = FALSE
)
}
\value{
\code{redist_smc} returns a \link{redist_plans} object containing the simulated
plans.
}
\description{
gSMC Redistricting Sampler
}
