% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_gsmc.R
\name{redist_gsmc}
\alias{redist_gsmc}
\title{gSMC Redistricting Sampler}
\usage{
redist_gsmc(
  state_map,
  M,
  counties = NULL,
  k_params = 6,
  resample = TRUE,
  runs = 1L,
  ncores = 0L,
  pop_temper = 0,
  verbose = FALSE,
  silent = FALSE,
  diagnostic_mode = FALSE
)
}
\arguments{
\item{k_params}{Either a single value to use as the splitting parameter for
every round or a vector of length N-1 where each value is the one to use for
a split.}
}
\value{
\code{redist_smc} returns a \link{redist_plans} object containing the simulated
plans.
}
\description{
gSMC Redistricting Sampler
}
