% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_mmss.R
\name{redist_mmss}
\alias{redist_mmss}
\title{Multi-Merge-Split MCMC Redistricting Sampler}
\usage{
redist_mmss(
  map,
  nsims,
  warmup = 0,
  thin = 1L,
  l = 3L,
  init_plan = NULL,
  chains = 1L,
  counties = NULL,
  compactness = 1,
  constraints = list(),
  adapt_k_thresh = 0.99,
  k = NULL,
  ncores = NULL,
  cl_type = "PSOCK",
  return_all = TRUE,
  init_name = NULL,
  silly_adj_fix = FALSE,
  verbose = FALSE,
  silent = FALSE
)
}
\arguments{
\item{map}{A \link{redist_map} object.}

\item{nsims}{The number of samples to draw, including warmup.}

\item{warmup}{The number of warmup samples to discard. Recommended to be at
least the first 20\% of samples, and in any case no less than around 100
samples, unless initializing from a random plan.}

\item{thin}{Save every \code{thin}-th sample. Defaults to no thinning (1).}

\item{l}{The number of districts to merge and re-split at each step.
Must be at least 2. Defaults to 3. Higher values allow larger moves but may
have lower acceptance rates.}

\item{init_plan}{The initial state of the map, provided as a single vector
to be shared across all chains, or a matrix with \code{chains} columns.
If not provided, will default to the reference map of the \code{map} object, or
if none exists, will sample a random initial state using \link{redist_smc}.
You can also request a random initial state for each chain by setting
\code{init_plan="sample"}.}

\item{chains}{The number of chains to run. Each chain will have
\code{nsims} draws. If \code{init_plan} is sampled, each chain will be initialized
with its own sampled plan. Defaults to 1.}

\item{counties}{A vector containing county (or other administrative or
geographic unit) labels for each unit, which may be integers ranging from 1
to the number of counties, or a factor or character vector.  If provided,
the algorithm will generate maps that tend to follow county lines.}

\item{compactness}{Controls the compactness of the generated districts, with
higher values preferring more compact districts. Must be nonnegative. See
the 'Details' section for more information, and computational
considerations.}

\item{constraints}{A \link{redist_constr} object or list of constraints.}

\item{adapt_k_thresh}{The threshold value used in the heuristic to select a
value \code{k_i} for each splitting iteration. Set to 0.9999 or 1 if
the algorithm does not appear to be sampling from the target distribution.
Must be between 0 and 1.}

\item{k}{The number of edges to consider cutting after drawing a spanning
tree. Should be selected automatically in nearly all cases.}

\item{ncores}{The number of parallel processes to run. Defaults to the
number of available cores, capped at the number of chains. Only used when
\code{chains > 1}.}

\item{cl_type}{The cluster type (see \code{\link[parallel:makeCluster]{parallel::makeCluster()}}). Safest is
\code{"PSOCK"}, but \code{"FORK"} may be appropriate in some settings.}

\item{return_all}{If \code{TRUE} return all sampled plans; otherwise, just return
the final plan from each chain.}

\item{init_name}{A name for the initial plan, or \code{FALSE} to not include
the initial plan in the output.}

\item{silly_adj_fix}{Heuristic for fixing weird inputs.}

\item{verbose}{Whether to print out intermediate information while sampling.}

\item{silent}{Whether to suppress all diagnostic information.}
}
\value{
A \link{redist_plans} object containing the simulated plans.
}
\description{
\code{redist_mmss} uses a Markov Chain Monte Carlo algorithm that generalizes
the merge-split sampler (Carter et al. 2019) to merge and re-split \code{l}
districts at each step. When \code{l=2}, this reduces to the standard merge-split
algorithm. When \code{l=ndists}, this functions similar to a much less efficient SMC.
Higher values of \code{l} allow larger rearrangements of the plan at each step,
improving mixing at the cost of a lower acceptance rate.
}
\details{
This function draws samples from a specific target measure, controlled by the
\code{map}, \code{compactness}, and \code{constraints} parameters.

Key to ensuring good performance is monitoring the acceptance rate, which
is reported at the sample level in the output.
Users should also check diagnostics of the sample by running
\code{summary.redist_plans()}.

Higher values of \code{compactness} sample more compact districts;
setting this parameter to 1 is computationally efficient and generates nicely
compact districts.
}
\examples{
data(fl25)
fl_map <- redist_map(fl25, ndists = 3, pop_tol = 0.1)
sampled <- redist_mmss(fl_map, 100, l = 3)
}
\references{
Carter, D., Herschlag, G., Hunter, Z., and Mattingly, J. (2019). A
merge-split proposal for reversible Monte Carlo Markov chain sampling of
redistricting plans. arXiv preprint arXiv:1911.01503.

McCartan, C., & Imai, K. (2023). Sequential Monte Carlo for Sampling
Balanced and Compact Redistricting Plans. \emph{Annals of Applied Statistics} 17(4).
Available at \doi{10.1214/23-AOAS1763}.

DeFord, D., Duchin, M., and Solomon, J. (2019). Recombination: A family of
Markov chains for redistricting. arXiv preprint arXiv:1911.05725.
}
\concept{simulate}
