% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_optimal_gsmc_ms.R
\name{redist_optimal_gsmc_ms}
\alias{redist_optimal_gsmc_ms}
\title{OptimalgSMC with merge split Redistricting Sampler}
\usage{
redist_optimal_gsmc_ms(
  state_map,
  M,
  counties = NULL,
  k_params = 6,
  split_district_only = FALSE,
  run_ms = FALSE,
  ms_freq = 7,
  ms_steps_multiplier = 1L,
  resample = TRUE,
  runs = 1L,
  ncores = 0L,
  multiprocess = TRUE,
  pop_temper = 0,
  init_region_ids_mat = NULL,
  init_dvals_mat = NULL,
  verbose = FALSE,
  silent = FALSE,
  diagnostic_mode = FALSE
)
}
\arguments{
\item{k_params}{Either a single value to use as the splitting parameter for
every round or a vector of length N-1 where each value is the one to use for
a split.}

\item{multiprocess}{Whether or not to launch multiple processes (sometimes
better to disable to avoid using too much memory.)}
}
\value{
\code{redist_smc} returns a \link{redist_plans} object containing the simulated
plans.
}
\description{
OptimalgSMC with merge split Redistricting Sampler
}
