% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_smc_tidy.R
\name{redist_quantile_trunc}
\alias{redist_quantile_trunc}
\title{Helper function to truncate importance weights}
\usage{
redist_quantile_trunc(x)
}
\arguments{
\item{x}{the weights}
}
\value{
numeric vector
}
\description{
Defined as \code{pmin(x, quantile(x, 1 - length(x)^(-0.5)))}
}
\examples{
redist_quantile_trunc(c(1,2,3,4))

}
