% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_shortburst.R
\name{scorer-combine}
\alias{scorer-combine}
\alias{combine_scorers}
\alias{cbind.redist_scorer}
\title{Combine scoring functions}
\usage{
combine_scorers(...)

\method{cbind}{redist_scorer}(..., deparse.level = 1)
}
\arguments{
\item{...}{a numeric or a \code{redist_scorer} function, from \code{\link{scorers}}}

\item{deparse.level}{As in \code{\link[=cbind]{cbind()}}.}
}
\value{
function of class redist_scorer. Will return a matrix with each
column containing every plan's scores for a particular scoring function.
}
\description{
\code{redist_scorer} functions may be combined together to optimize along multiple
dimensions. Rather than linearly combining multiple scorers to form a single
objective as with \link{scorer-arith}, these functions allow analysts to approximate
the Pareto frontier for a set of scorers.
}
\concept{prepare}
