% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_shortburst.R
\name{scorers}
\alias{scorers}
\alias{scorer_frackept}
\alias{scorer_max_pct}
\title{Scoring fuctions for `redist_shortburst`}
\usage{
scorer_frackept(map)

scorer_max_pct(map, group_pop, total_pop, k = 1)
}
\arguments{
\item{map}{A \code{\link{redist_map}} object.}

\item{group_pop}{A numeric vector with the population of the group for every precinct.}

\item{total_pop}{A numeric vector with the population for every precinct.}
}
\value{
A single numeric value, wherel larger values are better.
}
\description{
The output of these functions may be passed into `redist_shortburst()` as
`score_fn`.
}
\details{
Function details:
- `scorer_frackept` returns the fraction of edges kept in each district.
Higher values mean more compactness.
- `scorer_rank_pct` returns the `k`-th top group percentage across districts.
For example, if the group is Democratic voters and `k=3`, then the function
returns the 3rd-highest fraction of Democratic voters across all districts.
Can be used to target `k` VRA districts or partisan gerrymanders.
}
