% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_shortburst.R
\name{scorers}
\alias{scorers}
\alias{scorer_frac_kept}
\alias{scorer_group_pct}
\title{Scoring fuctions for \code{redist_shortburst}}
\usage{
scorer_frac_kept(map)

scorer_group_pct(map, group_pop, total_pop, k = 1)
}
\arguments{
\item{map}{A \code{\link{redist_map}} object.}

\item{group_pop}{A numeric vector with the population of the group for every precinct.}

\item{total_pop}{A numeric vector with the population for every precinct.}

\item{k}{the k-th from the top group fraction to return as the score.}
}
\value{
A single numeric value, wherel larger values are better.
}
\description{
The output of these functions may be passed into \code{redist_shortburst()} as
\code{score_fn}.
}
\details{
Function details:
\itemize{
\item \code{scorer_frac_kept} returns the fraction of edges kept in each district.
Higher values mean more compactness.
\item \code{scorer_group_pct} returns the \code{k}-th top group percentage across districts.
For example, if the group is Democratic voters and \code{k=3}, then the function
returns the 3rd-highest fraction of Democratic voters across all districts.
Can be used to target \code{k} VRA districts or partisan gerrymanders.
}
}
