% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_plans.R
\name{subset_sampled}
\alias{subset_sampled}
\alias{subset_ref}
\title{Subset to sampled or reference draws}
\usage{
subset_sampled(plans)

subset_ref(plans)
}
\arguments{
\item{plans}{the \code{redist_plans} object}
}
\value{
a \code{redist_plans} object, with only rows corresponding to
simulated (or reference) draws remaining.
}
\description{
Subset to sampled or reference draws
}
\concept{analyze}
