% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{summary.redist_plans}
\alias{summary.redist_plans}
\title{Diagnostic information on sampled plans}
\usage{
\method{summary}{redist_plans}(object, ...)
}
\arguments{
\item{object}{a [redist_plans] object}

\item{\dots}{additional arguments (ignored)}
}
\value{
A data frame containing diagnostic information, invisibly.
}
\description{
Prints diagnostic information, which varies by algorithm.
R-hat values are calculated after rank-normalization and folding.
For summary statistics that vary across districts, R-hat is calculated for
the first district only.
}
