% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_plans.R
\name{summary.redist_plans}
\alias{summary.redist_plans}
\title{Diagnostic information on sampled plans}
\usage{
\method{summary}{redist_plans}(object, ...)
}
\arguments{
\item{object}{a [redist_plans] object}

\item{\dots}{additional arguments (ignored)}
}
\value{
A data frame containing diagnostic information, invisibly.
}
\description{
Currently only supported for [redist_smc()] output.
}
