% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treedist_gsmc.R
\name{treedist_gsmc}
\alias{treedist_gsmc}
\title{gSMC Redistricting Sampler (O'Sullivan, McCartan and Imai ???)}
\usage{
treedist_gsmc(
  map,
  nsims,
  counties = NULL,
  compactness = 1,
  constraints = list(),
  runs = 1L,
  alg_name,
  split_district_only = FALSE,
  weight_type = "optimal",
  splitting_method = UNIF_VALID_EDGE_SPLITTING,
  splitting_method_params = NULL,
  ms_freq = 0,
  ms_steps_multiplier = 1L,
  run_ms = 0 < ms_freq && ms_freq <= ndists,
  merge_prob_type = "uniform",
  resample = TRUE,
  num_processes = 0L,
  num_threads_per_process = 0L,
  multiprocess = TRUE,
  pop_temper = 0,
  init_region_ids_mat = NULL,
  init_region_sizes_mat = NULL,
  custom_size_split_list = NULL,
  num_splitting_steps = NULL,
  ref_name = NULL,
  verbose = FALSE,
  silent = FALSE,
  diagnostic_mode = FALSE
)
}
\value{
\code{treedist_gsmc} returns a \link{redist_plans} object containing the simulated
plans.
}
\description{
\code{treedist_gsmc} uses a Sequential Monte Carlo algorithm (O'Sullivan, McCartan and Imai ???)
to generate representative samples of congressional or legislative
redistricting plans according to contiguity, population, compactness, and
administrative boundary constraints. It samples plans on the space of spanning
forests.
}
\details{
This function draws samples from a specific target measure controlled by
the \code{map} parameters.
}
