% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_alg_helpers.R
\name{validate_constraints}
\alias{validate_constraints}
\title{Checks that a constraint input is valid and returns it as a list ready for c++}
\usage{
validate_constraints(map, constraints)
}
\arguments{
\item{map}{A \code{\link[=redist_map]{redist_map()}} object.}

\item{constraints}{A \code{\link[=redist_constr]{redist_constr()}} object or a list containing
information on sampling constraints. See \link{constraints} for more information.}
}
\value{
A list of splitting parameters which is safe to pass into c++ code
}
\description{
Checks that the sampling space/splitting method combination given is valid
and that the splitting parameters are valid.
}
