% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redist_alg_helpers.R
\name{validate_sample_space_and_splitting_method}
\alias{validate_sample_space_and_splitting_method}
\title{Checks the sampling space and splitting method are valid}
\usage{
validate_sample_space_and_splitting_method(
  sampling_space,
  splitting_method,
  splitting_params,
  num_splitting_steps
)
}
\arguments{
\item{sampling_space}{The space to sample the plans on. This does not affect
the plans output by the function but the sample space used can have a large
impact on computational cost/runtime and convergence. Current spaces supported
right now are
\itemize{
\item graph_plan_space : graph partition space
\item spanning_forest_space : spanning forest space
\item linking_edge_space : linking edge forest space
}}
}
\value{
A list of splitting parameters which is safe to pass into c++ code
}
\description{
Checks that the sampling space/splitting method combination given is valid
and that the splitting parameters are valid.
}
